+++ docs/joss/paper.bib
@@ -159,7 +159,8 @@
   booktitle =	 {Air Pollution Modeling and its Application XVII},
   pages =	 {580--589},
   year =	 2007,
-  publisher =	 {Springer}
+  publisher =	 {Springer},
+  doi =          {10.1007/978-0-387-68854-1_62}
 }
 
 @Article{liao24,
@@ -199,7 +200,8 @@
   year =	 2012,
   volume =	 200,
   series =	 {Geophysical Monograph Series},
-  address =	 {Washington DC}
+  address =	 {Washington DC},
+  doi =          {10.1029/GM200}
 }
 
 @InProceedings{liu20,
@@ -217,7 +219,7 @@
   publisher =	 "Springer International Publishing",
   address =	 "Cham",
   pages =	 "60--73",
-  doi =		 "https://doi.org/10.1007/978-3-030-50420-5_5"
+  doi =		 "10.1007/978-3-030-50420-5_5"
 }
 
 @Article{liu23,
@@ -306,7 +308,7 @@
   number =	 12,
   journal =	 "Bulletin of the American Meteorological Society",
   pages =	 "2059 - 2077",
-  doi =		 "https://doi.org/10.1175/BAMS-D-14-00110.1"
+  doi =		 "10.1175/BAMS-D-14-00110.1"
 }
 
 @Article{stohl05,
