### 2025-04-09 3c12555
Added halos for 2d fields
