+++ src/mptrac.c
@@ -8270,7 +8270,22 @@ int read_met_nc_2d_par(
       ERRMSG("Meteo data layout not implemented for packed netCDF files!");
 
     /* Copy and check data... */
-#pragma omp parallel for default(shared) num_threads(12)
+    #pragma omp parallel for default(shared) num_threads(12)
+    for (int ix = 0; ix < met->nx; ix++)
+      for (int iy = 0; iy < met->ny; iy++) {
+	      if (init)
+	        dest[ix][iy] = 0;
+	short aux = help[ARRAY_2D(iy, ix, met->nx)];
+	if ((fillval == 0 || aux != fillval)
+	    && (missval == 0 || aux != missval)
+	    && fabsf(aux * scalfac + offset) < 1e14f)
+	  dest[ix][iy] += scl * (aux * scalfac + offset);
+	else
+	  dest[ix][iy] = NAN;
+      }
+
+    /* Copy and check data... */
+    #pragma omp parallel for default(shared) num_threads(12)
     for (int ix = 0; ix < met->nx; ix++)
       for (int iy = 0; iy < met->ny; iy++) {
 	if (init)
@@ -8322,35 +8337,67 @@ int read_met_nc_2d_par(
     help_domain_count[1] = met->domain_count[2];
     help_domain_count[2] = met->domain_count[3];
     
-    //ALLOC(help, float, (int) met->domain_count[2] * (int) met->domain_count[3]);
+    // (int) met->domain_count[1] * 
     ALLOC(help, float, 
-    (int) met->domain_count[1] * (int) met->domain_count[2] *
+    (int) met->domain_count[2] *
     (int) met->domain_count[3]
     );
 
     /* Read data... */
-    //printf("[Debugging] start %ld:%ld:%ld:%ld, count %ld:%ld:%ld:%ld \n",
-    //met->domain_start[0], met->domain_start[1], met->domain_start[2], met->domain_start[3],
-    //met->domain_count[0], met->domain_count[1], met->domain_count[2], met->domain_count[3]);
     NC(nc_get_vara_float(ncid, varid, help_domain_start, help_domain_count, help));
+
+    /* Read halos at boundaries...*/
+    size_t help_halo_bnd_start[3];
+    size_t help_halo_bnd_count[3];
+
+    //help_domain_start[0] = 0;
+    help_halo_bnd_start[0] = 0; 
+    help_halo_bnd_start[1] = met->halo_bnd_start[2];
+    help_halo_bnd_start[2] = met->halo_bnd_start[3];
+   
+    //help_domain_count[0] = 1;
+    help_halo_bnd_count[0] = 1;
+    help_halo_bnd_count[1] = met->halo_bnd_count[2];
+    help_halo_bnd_count[2] = met->halo_bnd_count[3];
+
+    float* help_halo;
+    ALLOC(help_halo, float, met->halo_bnd_count[2]*met->halo_bnd_count[3])     
+    NC(nc_get_vara_float(ncid, varid,  help_halo_bnd_start, help_halo_bnd_count, help_halo));
 
     /* Check meteo data layout... */
     if (ctl->met_convention == 0) {
 
+      printf("Copy 2d data 1\n");
       /* Copy and check data (ordering: lat, lon)... */
 #pragma omp parallel for default(shared) num_threads(12)
-      for (int ix = 0; ix < met->nx; ix++)
-	for (int iy = 0; iy < met->ny; iy++) {
-	  if (init)
-	    dest[ix][iy] = 0;
-	  float aux = help[ARRAY_2D(iy, ix, met->nx)];
-	  if ((fillval == 0 || aux != fillval)
+  for (int ix = 0; ix < (int) help_domain_count[2]; ix++)
+	  for (int iy = 0; iy < (int) help_domain_count[1]; iy++) {
[…]
