### 2025-04-11 0885d69
Clean-up of halos for 2d fields
