+++ src/mptrac.c
@@ -7332,7 +7332,7 @@ void read_met_grid(
   /* Get the MPI information... */
   int rank, size;
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
-  MPI_Comm_rank(MPI_COMM_WORLD, &size);
+  MPI_Comm_size(MPI_COMM_WORLD, &size);
 
   /* Check for edge cases... */
   bool left = (rank <= ctl->dd_domains_meridional-1);
@@ -7401,22 +7401,23 @@ void read_met_grid(
   double lon_shift = 0;
   if ( left || right ) {
 
-      met->nx = met->nx + ctl->dd_halos_size;
+    met->nx = met->nx + ctl->dd_halos_size;
     
-      met->halo_bnd_start[0] = 0;
-      met->halo_bnd_start[1] = 0;
-      met->halo_bnd_start[3] = (size_t)(left ? (met->nx_glob - ctl->dd_halos_size) : (0)); //x
-      met->halo_bnd_start[2] = met->domain_start[2]; //y
+    met->halo_bnd_start[0] = 0;
+    met->halo_bnd_start[1] = 0;
+    met->halo_bnd_start[3] = (size_t)(left ? (met->nx_glob - ctl->dd_halos_size) : (0)); //x
+    met->halo_bnd_start[2] = met->domain_start[2]; //y
       
-      met->halo_bnd_count[0] = 1;
-      met->halo_bnd_count[1] = (size_t) met->np;
-      met->halo_bnd_count[3] = (size_t) ctl->dd_halos_size;
-      met->halo_bnd_count[2] = (size_t) met->ny + (size_t) ctl->dd_halos_size * ((top || bottom) ?  1 : 2 ); 
+    met->halo_bnd_count[0] = 1;
+    met->halo_bnd_count[1] = (size_t) met->np;
+    met->halo_bnd_count[3] = (size_t) ctl->dd_halos_size;
+    met->halo_bnd_count[2] = (size_t) met->ny + (size_t) ctl->dd_halos_size * ((top || bottom) ?  1 : 2 ); 
   
-      met->halo_offset_start =  (left ? (int) met->halo_bnd_count[3] : 0);
-      met->halo_offset_end = (left ? 0 : (int) met->domain_count[3]);
-      lon_shift = (left ? -360 : 360);
-  }  
+    met->halo_offset_start =  (left ? (int) met->halo_bnd_count[3] : 0);
+    met->halo_offset_end = (left ? 0 : (int) met->domain_count[3]);
+    lon_shift = (left ? -360 : 360);
+
+  }
 
   /* Get the range of the entire meteodata... */
   double lon_range = met->lon[ met->nx_glob - 1 ] - met->lon[0];
@@ -7466,12 +7467,17 @@ void read_met_grid(
   LOG(2, " %d Domain latitudes: %g, %g ... %g deg", rank,
       met->lat[0], met->lat[1], met->lat[met->ny - 1]);
 
-  LOG(2, "Max dim sizes: %d-%d-%d : Met: %d-%d-%d : Halo Bnd: %d-%d-%d : Domain :  %d-%d-%d", EX, EY, EP, met->nx, met->ny, met->np, 
-    (int) met->halo_bnd_count[3], (int) met->halo_bnd_count[2], (int) met->halo_bnd_count[1],
-    (int) met->domain_count[3], (int) met->domain_count[2], (int) met->domain_count[1]);
-  LOG(2, "Domain starts %ld-%ld-%ld : Halo start : %ld-%ld-%ld : Offsets: %d-%d",met->domain_start[3], met->domain_start[2], met->domain_start[1],
-  met->halo_bnd_start[3], met->halo_bnd_start[2], met->halo_bnd_start[1], met->halo_offset_start, met->halo_offset_end);
-
[…]
