+++ src/mptrac.c
@@ -4646,7 +4646,7 @@ void mptrac_read_ctl(
   ctl->qnt_Csf6 = -1;
   ctl->qnt_aoa = -1;
   ctl->qnt_destination = -1;
-  ctl->qnt_domain = -1;
+  ctl->qnt_subdomain = -1;
 
   /* Read quantities... */
   ctl->nq = (int) scan_ctl(filename, argc, argv, "NQ", -1, "0", NULL);
@@ -4764,8 +4764,8 @@ void mptrac_read_ctl(
       SET_QNT(qnt_Cn2o, "Cn2o", "N2O volume mixing ratio", "ppv")
       SET_QNT(qnt_Csf6, "Csf6", "SF6 volume mixing ratio", "ppv")
       SET_QNT(qnt_aoa, "aoa", "age of air", "s")
-      SET_QNT(qnt_destination, "destination", "domain index of destination", "-")
-      SET_QNT(qnt_domain, "domain", "current domain index", "-")
+      SET_QNT(qnt_destination, "destination", "subdomain index of destination", "-")
+      SET_QNT(qnt_subdomain, "subdomain", "current subdomain index", "-")
       scan_ctl(filename, argc, argv, "QNT_UNIT", iq, "", ctl->qnt_unit[iq]);
   }
 
@@ -5382,10 +5382,10 @@ void mptrac_read_ctl(
     (int) scan_ctl(filename, argc, argv, "VTK_SPHERE", -1, "0", NULL);
     
   /* Controle of domain decomposition... */
-  ctl->dd_domains_meridional =
-    (int) scan_ctl(filename, argc, argv, "DD_DOMAINS_MERIDIONAL", -1, "1", NULL);
-  ctl->dd_domains_zonal =
-    (int) scan_ctl(filename, argc, argv, "DD_DOMAINS_ZONAL", -1, "1", NULL);
+  ctl->dd_subdomains_meridional =
+    (int) scan_ctl(filename, argc, argv, "DD_SUBDOMAINS_MERIDIONAL", -1, "1", NULL);
+  ctl->dd_subdomains_zonal =
+    (int) scan_ctl(filename, argc, argv, "DD_SUBDOMAINS_ZONAL", -1, "1", NULL);
   ctl->dd_nbr_neighbours =
     (int) scan_ctl(filename, argc, argv, "DD_NBR_NEIGHBOURS", -1, "8", NULL);
   ctl->dd_halos_size =
@@ -5570,7 +5570,7 @@ void mptrac_run_timestep(
   //}
 
   /* Domain decomposition... */
-  if (ctl->dd_domains_meridional*ctl->dd_domains_zonal > 1) {
+  if (ctl->dd_subdomains_meridional*ctl->dd_subdomains_zonal > 1) {
     //module_dd()
     
     /* Initialize particles locally... */
@@ -5581,23 +5581,23 @@ void mptrac_run_timestep(
     int rank;
     MPI_Barrier(MPI_COMM_WORLD);
     MPI_Comm_rank(MPI_COMM_WORLD, &rank);
-    int size = ctl->dd_domains_meridional*ctl->dd_domains_zonal;
+    int size = ctl->dd_subdomains_meridional*ctl->dd_subdomains_zonal;
 
     LOG(1, "Starting module_dd with %d particles on rank %d", atm->np, rank);
         
     /* Register the MPI_Particle data type... */
-    LOG(1, "Register the MPI_Particle data type.")
+    LOG(1, "Register the MPI_Particle data type.");
     MPI_Datatype MPI_Particle;
     dd_register_MPI_type_particle(&MPI_Particle);
     
-    /* Define communication destinations ... */
-    LOG(1, "Define communication destinations.")
[…]
