+++ src/mptrac.h
@@ -274,12 +274,12 @@
 #define EP 150
 #endif
 
-/*! Maximum number of domain longitudes for meteo data. */
+/*! Maximum number of subdomain longitudes for meteo data. */
 #ifndef EX
 #define EX 1202
 #endif
 
-/*! Maximum number of domain latitudes for meteo data. */
+/*! Maximum number of subdomain latitudes for meteo data. */
 #ifndef EY
 #define EY 602
 #endif
@@ -297,6 +297,11 @@
 /*! Maximum number of global latitudes for meteo data. */
 #ifndef EY_GLOB
 #define EY_GLOB 602
+#endif
+
+/*! Maximum number of neighbours to communicate with. */
+#ifndef NNMAX
+#define NNMAX  26
 #endif
 
 /*! Maximum number of data points for ensemble analysis. */
@@ -3216,11 +3221,11 @@ typedef struct {
   /*! Spherical projection for VTK data (0=no, 1=yes). */
   int vtk_sphere;
   
-  /*! Zonal domain number. */
-  int dd_domains_zonal;
+  /*! Zonal subdomain number. */
+  int dd_subdomains_zonal;
   
-  /*! Meridional domain number. */
-  int dd_domains_meridional;
+  /*! Meridional subdomain number. */
+  int dd_subdomains_meridional;
   
   /*! Number of neighbours to communicate with. */
   int dd_nbr_neighbours;
@@ -3228,8 +3233,8 @@ typedef struct {
   /*! Size of halos given in grid-points. */
   int dd_halos_size;
 
-  /*! Quantity array index for the current domain. */
-  int qnt_domain;
+  /*! Quantity array index for the current subdomain. */
+  int qnt_subdomain;
 
   /*! Quantity array index for the destination domain */
   int qnt_destination;
@@ -3677,13 +3682,13 @@ typedef struct {
   
   /*! More grid information... */
   // TODO: Or this will be integrated to a grid_t ...
-  double domain_lon_max;
-  double domain_lon_min;
-  double domain_lat_max;
-  double domain_lat_min;
+  double subdomain_lon_max;
+  double subdomain_lon_min;
+  double subdomain_lat_max;
+  double subdomain_lat_min;
   
-  size_t domain_start[4];
-  size_t domain_count[4];
+  size_t subdomain_start[4];
+  size_t subdomain_count[4];
 
   size_t halo_bnd_start[4];
   size_t halo_bnd_count[4];
@@ -3695,7 +3700,13 @@ typedef struct {
   int nx_glob;
   int ny_glob;
   int np_glob;
+
+  /* TODO: Number of neighbours shif tto here?*/
   
+  /* Neighbour ranks of the subdomain. */
+  /* TODO: use this instead of inline neighbours array...*/
+  int neighbours[NNMAX];
+
 } met_t;
 
 /* ------------------------------------------------------------
@@ -8868,7 +8879,7 @@ void particles2atm(
 /**
  * @brief Creates an MPI datatype for particle structures.
  *
- * The `dd_destination_get_rect` function defines an MPI datatype for
+ * The `dd_neighbour_get_rect` function defines an MPI datatype for
  * particle structures (`particle_t`). This datatype is used to describe
  * the layout of particle data in memory, enabling efficient communication
  * of particle data in MPI programs.
@@ -8892,33 +8903,33 @@ void  dd_register_MPI_type_particle(
   MPI_Datatype* MPI_Particle);
   
 /**
- * @brief Determines destination ranks for data distribution in a domain decomposition setup.
+ * @brief Determines neighbour ranks for data distribution in a domain decomposition setup.
  *
- * The `dd_get_destinations` function calculates and assigns destination ranks for data
+ * The `dd_get_neighbours` function calculates and assigns neighbour ranks for data
[…]
