### 2025-04-11 009c27a
Made read_met_periodic and timestep availiable for non dd cases again...
