### 2025-04-17 #8031 @kthyng
@lars2015

* It looks like you are missing formal docs. Can you expand your readme out into something like readthedocs to demonstrate your examples and contain other useful info?
* Also please shorten your paper by 200-300 words.
