+++ docs/joss/paper.md
@@ -79,40 +79,38 @@ bibliography: paper.bib
 
 # Summary
 
-The Massive-Parallel Trajectory Calculations (MPTRAC) Lagrangian transport model [@hoffmann16; @hoffmann22] simulates the movement, dispersion, and chemical transformation of aerosol particles and trace gases in the atmosphere. Unlike grid-based (Eulerian) models, MPTRAC tracks individual air parcel trajectories, providing high-resolution simulations of atmospheric transport processes. The Lagrangian approach precisely models dispersion, mixing, chemistry, and deposition along the trajectories. Trajectory and advection calculations in MPTRAC require meteorological fields from data assimilation systems and forecast models. MPTRAC has been used in various applications, particularly for tracking volcanic plumes in the free troposphere and stratosphere. Optimized for high-performance computing (HPC) and graphics processing unit (GPU) systems, MPTRAC efficiently handles large-scale, high-resolution simulations, making it a powerful tool for atmospheric research, source attribution, and emergency response.
+The Massive-Parallel Trajectory Calculations (MPTRAC) Lagrangian transport model [@hoffmann16; @hoffmann22] simulates the movement, dispersion, and chemical transformation of aerosol particles and trace gases in the atmosphere. Unlike grid-based models, MPTRAC tracks individual air parcels, offering high-resolution insights into transport, mixing, chemistry, and deposition. Trajectory calculations are driven by meteorological fields from data assimilation systems and forecast models. MPTRAC has been applied in various contexts, especially for tracking volcanic plumes in the troposphere and stratosphere. Optimized for high-performance computing (HPC) and graphics processing unit (GPU) systems, MPTRAC efficiently supports large-scale, high-resolution simulations, making it a powerful tool in atmospheric research.
 
 # Statement of need
 
-Accurate atmospheric transport modeling is essential for applications such as air quality assessments, climate studies, and public health protection. Traditional Eulerian models often struggle with fine-scale resolution, limiting their ability to capture complex transport and mixing processes. Lagrangian models, such as MPTRAC and others [@mckenna02; @lin03; @stohl05; @jones07; @stein15; @pisso19], address this limitation by tracking individual air parcels, offering a more detailed and flexible representation of dispersion, chemical reactions, and deposition of atmospheric aerosol particles and trace gases.
+Accurate atmospheric transport modeling is essential for applications such as air quality assessments, climate studies, and public health protection. Traditional Eulerian models often struggle with fine-scale resolution, limiting their ability to capture complex transport and mixing processes. Lagrangian models, such as MPTRAC and others [@mckenna02; @lin03; @stohl05; @jones07; @stein15; @pisso19], address this limitation by tracking individual air parcels, offering fine-grained representations of atmospheric processes.
 
-MPTRAC is primarily designed for researchers to study atmospheric transport processes, but it also supports operational applications requiring high-precision atmospheric transport simulations. Its ability to model long-range transport and chemical transformations of atmospheric constituents enables studies of pollution dispersion, tracking of hazardous emissions, and environmental impact forecasting. In contrast to many other Lagrangian models, MPTRAC is optimized for HPC and GPU systems, enabling fast, large-scale simulations that support real-time decision-making and scientific analysis.
[…]
