### 2025-04-20 5967e80
Update mkdocs manual.
