### 2025-04-20 7f07caa
Update mkdocs manual.
