+++ docs/manual/tropopause-data.md
@@ -1,14 +1,24 @@
 # Tropopause data
 
-This page provides information on the methods used to calculate and extract tropopause data with MPTRAC.
+This page provides information on the methods used to calculate and
+extract tropopause data with MPTRAC.
 
 ## Determination of the tropopause
 
-In general, the tropopause data in MPTRAC are calculated from the given meteorological input data, i.e, the data are not read from the meteo data files.
+In general, the tropopause data in MPTRAC are calculated from the
+given meteorological input data, i.e, the data are not read from the
+meteo data files.
 
-For the determination of the height level of the tropopause, the meteorological data are first interpolated to a fine vertical grid by means of cubic spline interpolation. The tropopause level is estimated based on the interpolated data. This approach improves the vertical resolution of the tropopause data. However, the vertical resolution of the meteorological input data will also play an important role.
+For the determination of the height level of the tropopause, the
+meteorological data are first interpolated to a fine vertical grid by
+means of cubic spline interpolation. The tropopause level is estimated
+based on the interpolated data. This approach improves the vertical
+resolution of the tropopause data. However, the vertical resolution of
+the meteorological input data will also play an important role.
 
-The method used to extract the height level of the tropopause is selected by means of the parameter MET_TROPO. The following options to determine the tropopause level have been implemented:
+The method used to extract the height level of the tropopause is
+selected by means of the parameter `MET_TROPO`. The following options to
+determine the tropopause level have been implemented:
 
 | MET_TROPO   | Description       |
 | ----------- | ----------------- |
@@ -21,32 +31,36 @@ The method used to extract the height level of the tropopause is selected by mea
 
 ## Tropopause data along trajectories
 
-The trac tool of MPTRAC can extract the tropopause pressure along the trajectories.
-
-The tropopause pressure needs to be selected as quantity "pt", for example:
+The trac tool of MPTRAC can extract the tropopause data along the
+trajectories. For example, set the following quantities to
+obtain tropopause pressure, height, and temperature:
 
 ```
-    NQ = 1
-    QNT_NAME[0] = pt
+    NQ = 3
+    QNT_NAME[0] = pt   # tropopause pressure [hPa]
+    QNT_NAME[1] = zt   # tropopause geootential height [km]
+    QNT_NAME[2] = tt   # tropopause temperature [K]
 ```
-
-The default unit is "hPa".
 
 ## Tropopause data files
 
-MPTRAC provides the tool "tropo" that generates tropopause data files from given meteorological input data.
+MPTRAC provides the tool `tropo` that generates tropopause data files
+from given meteorological input data.
 
-The tropopause data are provided as netCDF files. Each data file may contain one or more time steps of the reanalysis.
+The tropopause data are provided as netCDF files. Each data file may
+contain one or more time steps of the reanalysis.
 
-The data files provide geopotential height, pressure, temperature, and water vapor volume mixing ratios for the cold point, WMO first and second tropopause, and the dynamical tropopause.
+The data files provide geopotential height, pressure, temperature, and
+water vapor volume mixing ratios for the cold point, WMO first and
+second tropopause, and the dynamical tropopause.
 
-The tool "tropo_sample" can be used to extract tropopause data from the netCDF files created by "tropo" for a set of locations specified as a particle data file (atm.tab file).
+The tool `tropo_sample` can be used to extract tropopause data from
+the netCDF files created by `tropo` for a set of locations specified
+as a particle data file (`atm.tab` file).
 
[…]

[…]
