### 2025-04-20 ed023b5
Update mkdocs manual.
