### 2025-04-20 68ce874
Update mkdocs manual.
