+++ docs/manual/meteorological-input-data.md
@@ -41,19 +41,9 @@ of meteo data netCDF files used for MPTRAC.
 ### Overview
 
 As an example, we discuss the use of
-[ERA-Interim](https://www.ecmwf.int/en/forecasts/datasets/reanalysis-datasets/era-interim)
-and
 [ERA5](https://www.ecmwf.int/en/forecasts/datasets/reanalysis-datasets/era5)
 reanalysis data provided by the European Centre for Medium-Range
 Weather Forecasts (ECMWF) to run the MPTRAC model.
-
-ERA-Interim is a global atmospheric reanalysis that is available from
-1 January 1979 to 31 August 2019. The data assimilation system used to
-produce ERA-Interim is based on a 2006 release of the IFS
-(Cy31r2). The system includes a 4-dimensional variational analysis
-(4D-Var) with a 12-hour analysis window. The spatial resolution of the
-data set is approximately 80 km (T255 spectral grid) on 60 levels in
-the vertical from the surface up to 0.1 hPa.
 
 ERA5 provides hourly estimates of a large number of atmospheric, land
 and oceanic climate variables from 1950 to present. The data cover the
@@ -64,13 +54,7 @@ at reduced spatial and temporal resolutions.
 
 ### Download of ECMWF data
 
-The first step is to download the ERA-Interim data from [ECMWF's data
-archive](https://apps.ecmwf.int/datasets/data/interim-full-daily). It
-is recommended to download the data in grib file format at 0.75°x0.75°
-horizontal resolution, on all 60 model levels, and at 6-hourly time
-intervals.
-
-Alternatively, download ERA5 data from the
+The first step is to download the ERA5 data from the
 [Copernicus Climate Change Service (C3S) Climate Date Store](https://cds.climate.copernicus.eu/#!/search?text=ERA5&type=dataset).
 It is recommended to download the data in grib file format at 0.3°x0.3°
 horizontal resolution, on all 137 model levels, and at hourly time
@@ -78,17 +62,20 @@ intervals.
 
 ### Interpolation to pressure levels
 
-The second step is to interpolate the ERA-Interim data from model
-levels to pressure levels. Here, a set of target pressure levels needs
-to be selected. It is recommended to use a set of 60 pressure levels
-that matches the vertical sampling of the model level data (as defined
+The second step is to interpolate the ERA5 data from model levels to
+pressure levels. Here, a set of target pressure levels needs to be
+selected. It is recommended to use a set of 137 pressure levels that
+matches the vertical sampling of the IFS model level data as defined
 by a fixed surface pressure of 1013.25 hPa and
-[ECMWF's a and b level coefficients](https://www.ecmwf.int/en/forecasts/documentation-and-support/60-model-levels)).
+[ECMWF's a and b level coefficients](https://www.ecmwf.int/en/forecasts/documentation-and-support/137-model-levels).
 
-For ERA5 a different set of [a and be level coefficients](https://www.ecmwf.int/en/forecasts/documentation-and-support/137-model-levels)
-should be considered. Potentially, the vertical sampling of the target
-pressure levels in the lower troposphere and the number of levels in
-the mesosphere can be reduced to reduce the amount of data.
+Potentially, the vertical sampling of the target pressure levels in
+the lower troposphere and the number of levels in the mesosphere can
+be reduced to reduce the amount of data.  However, it is highly
[…]
