### 2025-04-21 f28222d
Update README.md and mkdocs manual.
