+++ docs/manual/model-physics.md
@@ -2,7 +2,9 @@
 
 ## Advection
 
-The advection of an air parcel, i.\,e., the position $\vec{x}(t)$ at time $t$ for a given wind and velocity field $\vec{v}(\vec{x}, t)$, is given by the trajectory equation,
+The advection of an air parcel, i.\,e., the position $\vec{x}(t)$ at
+time $t$ for a given wind and velocity field $\vec{v}(\vec{x}, t)$, is
+given by the trajectory equation,
 
 $$
 \begin{align}
@@ -10,11 +12,18 @@ $$
 \end{align}
 $$
 
-Here, the position $\vec{x}$ is provided in a meteorological coordinate system, $\vec{x}=(\lambda,\phi,p)$, with longitude $\lambda$, latitude $\phi$, and a vertical coordinate (pressure $p$ or zeta $\zeta$). The velocity vector $\vec{v}=(u,v,\omega)$ is composed of the zonal wind component $u$, the meridional wind component $v$, and the vertical velocity ($\omega$ or $\dot{\zeta}$), respectively. 
+Here, the position $\vec{x}$ is provided in a meteorological
+coordinate system, $\vec{x}=(\lambda,\phi,p)$, with longitude
+$\lambda$, latitude $\phi$, and a vertical coordinate (pressure $p$ or
+zeta $\zeta$). The velocity vector $\vec{v}=(u,v,\omega)$ is composed
+of the zonal wind component $u$, the meridional wind component $v$,
+and the vertical velocity ($\omega$ or $\dot{\zeta}$), respectively.
 
 #### Integration Method
 
-Based on its accuracy and computational efficiency for trajectory calculations (Rößler et al., 2018), per default we apply the explicit midpoint method to solve the trajectory equation,
+Based on its accuracy and computational efficiency for trajectory
+calculations (Rößler et al., 2018), per default we apply the explicit
+midpoint method to solve the trajectory equation,
 
 $$
 \begin{align}
@@ -22,7 +31,9 @@ $$
 \end{align}
 $$
 
-As an alternative, the classical 4th-order Runge-Kutta method can be employed for the simulation, which has higher accuracy at higher computational costs. The method is defined by the equations:
+As an alternative, the classical 4th-order Runge-Kutta method can be
+employed for the simulation, which has higher accuracy at higher
+computational costs. The method is defined by the equations:
 
 $$
 \begin{align}
@@ -39,19 +50,29 @@ $$
 \end{align}
 $$
 
-Finally, MPTRAC can use the Euler method as well. The Euler method has low accuracy but is the fastest option available.  
+Finally, MPTRAC can use the Euler method as well. The Euler method has
+low accuracy but is the fastest option available.
 
-The integration method is selected with the parameter ADVECT.
+The integration method is selected with the parameter `ADVECT`.
 
-| ADVECT | Integration Method | 
-|---|---|
-|  1 | explicit Euler   | 
-|  2 (default) | mid-point  |
-|  4 |  classical 4th order Runge-Kutta  |
+| ADVECT      | Integration Method                | 
+|-------------|-----------------------------------|
+| 1           | explicit Euler                    | 
+| 2 (default) | mid-point                         |
+| 4           |  classical 4th order Runge-Kutta  |
 
 #### Vertical coordinate and diabatic/kinematic velocities
 
[…]
