### 2025-04-21 405a943
Update README.md and mkdocs manual.
