### 2025-04-21 206e34a
Update mkdocs manual.
