### 2025-04-21 e9dc0ed
Update JOSS paper.
