+++ docs/joss/paper.bib
@@ -151,6 +151,16 @@
   DOI =		 {10.5194/gmd-17-4077-2024}
 }
 
+@misc{hoffmann25,
+  author =	 {Hoffmann, Lars},
+  title =	 {MPTRAC Convective Transport Animation},
+  month =	 apr,
+  year =	 2025,
+  publisher =	 {Zenodo},
+  version =	 {1.0},
+  doi =		 {10.5281/zenodo.15256283}
+}
+
 @incollection{jones07,
   title =	 {The {UK Met Office's} next-generation atmospheric
                   dispersion model, {NAME III}},
@@ -160,7 +170,7 @@
   pages =	 {580--589},
   year =	 2007,
   publisher =	 {Springer},
-  doi =          {10.1007/978-0-387-68854-1_62}
+  doi =		 {10.1007/978-0-387-68854-1_62}
 }
 
 @Article{liao24,
@@ -201,7 +211,7 @@
   volume =	 200,
   series =	 {Geophysical Monograph Series},
   address =	 {Washington DC},
-  doi =          {10.1029/GM200}
+  doi =		 {10.1029/GM200}
 }
 
 @InProceedings{liu20,
