+++ docs/joss/paper.bib
@@ -152,7 +152,7 @@
 }
 
 @misc{hoffmann25,
-  author =	 {Hoffmann, Lars},
+  author =	 {Hoffmann, L.},
   title =	 {MPTRAC Convective Transport Animation},
   month =	 apr,
   year =	 2025,
@@ -174,8 +174,8 @@
 }
 
 @Article{liao24,
-  AUTHOR =	 {Liao, Ye and Deng, Xuying and Huang, Mingming and
-                  Liu, Mingzhao and Yi, Jia and Hoffmann, Lars},
+  AUTHOR =	 {Liao, Y. and Deng, X. and Huang, M. and Liu, M. and
+                  Yi, J. and Hoffmann, L.},
   TITLE =	 {Tracking Carbon Dioxide with {Lagrangian} Transport
                   Simulations: Case Study of {Canadian} Forest Fires
                   in May 2021},
@@ -215,8 +215,8 @@
 }
 
 @InProceedings{liu20,
-  author =	 "Liu, Mingzhao and Huang, Yaopeng and Hoffmann, Lars
-                  and Huang, Chunyan and Chen, Pin and Heng, Yi",
+  author =	 "Liu, M. and Huang, Y. and Hoffmann, L. and Huang,
+                  C. and Chen, P. and Heng, Y.",
   editor =	 "Krzhizhanovskaya, Valeria V.  and Z{\'a}vodszky,
                   G{\'a}bor and Lees, Michael H.  and Dongarra, Jack
                   J.  and Sloot, Peter M. A.  and Brissos, S{\'e}rgio
@@ -262,8 +262,7 @@
 }
 
 @Article{mishra22,
-  AUTHOR =	 {Mishra, Manoj Kumar and Hoffmann, Lars and
-                  Thapliyal, Pradeep Kumar},
+  AUTHOR =	 {Mishra, M. K. and Hoffmann, L. and Thapliyal, P. K.},
   TITLE =	 {Investigations on the Global Spread of the {Hunga
                   Tonga-Hunga Ha’apai} Volcanic Eruption Using
                   Space-Based Observations and {Lagrangian} Transport
