### 2025-04-22 f16d980
Update mkdocs manual.
