+++ docs/manual/fortran-wrapper.md
@@ -1,18 +1,27 @@
 # Fortran wrapper
 
-MPTRAC has been equiped with a wrapper to access the C functions from Fortran. The project started as a coding sprint in the [natESM project LAGOOn](https://www.nat-esm.de/services/support-through-sprints/documentation).
+MPTRAC has been equiped with a wrapper to access the C functions from
+Fortran. The project started as a coding sprint in the
+[natESM project LAGOOn](https://www.nat-esm.de/services/support-through-sprints/documentation).
 
-To integrate multi language programming with Fortran and C in a single program the [ISO_C_BINDINGS](https://gcc.gnu.org/onlinedocs/gfortran/ISO_005fC_005fBINDING.html) module is since Fortran 2003 the standard approach that facilitates interoperability between the two languages. In Fortran, the compiler is informed with the BIND(C) attribute that a symbol shall be interoperable with C.
+To integrate multi language programming with Fortran and C in a single
+program the [ISO_C_BINDINGS](https://gcc.gnu.org/onlinedocs/gfortran/ISO_005fC_005fBINDING.html)
+module is since Fortran 2003 the standard approach that facilitates
+interoperability between the two languages. In Fortran, the compiler
+is informed with the BIND(C) attribute that a symbol shall be
+interoperable with C.
 
 ## Compilation and testing
 
-To compile the Fortran wrapper, navigate to the `src` directory and run the following command:
+To compile the Fortran wrapper, navigate to the `src` directory and
+run the following command:
 
 ```
 $ make wrapper
 ```
 
-To test the compiled wrapper, use the following command in the same directory:
+To test the compiled wrapper, use the following command in the same
+directory:
 
 ```
 $ make wrapper_test
@@ -22,7 +31,9 @@ $ make wrapper_test
 
 ### trac_fortran.f90
 
-This program calculates trajectories based on MPTRAC's module_advect function. It is the Fortran counterpart to trac.c. The program is called with at least four (4) arguments:
+This program calculates trajectories based on MPTRAC's module_advect
+function. It is the Fortran counterpart to trac.c. The program is
+called with at least four (4) arguments:
 
 ```
 # calling trac_fortran
@@ -31,13 +42,18 @@ $ ./trac_fortran  <dirlist> <ctl> <atm_in> <metbase>
 
 The required arguments are:
 
-* dirlist: A file containing directories to be processed. Each directory has to have an own control parameter file and a starting point file.
+* `dirlist`: A file containing directories to be processed. Each
+  directory has to have an own control parameter file and a starting
+  point file.
 
-* ctl: In the control parameter file the configuration parameters can be set.
+* `ctl`: In the control parameter file the configuration parameters can
+  be set.
 
-* atm_in: The starting point file contains a list of starting points for the trajectory calculation.
+* `atm_in`: The starting point file contains a list of starting points
+  for the trajectory calculation.
 
-* metbase: Here, the path to the meteorological data files and their basename shall be given.
+* `metbase`: Here, the path to the meteorological data files and their
+  basename shall be given.
 
 Example from the repository:
 
@@ -46,7 +62,9 @@ Example from the repository:
 $ ./trac_fortran data/dirlist trac.ctl atm_split.tab meteo/ei
 ```
 
-In addition it is possible to append control parameters, giving the flag name first, followed by the parameter, all separated by a space. Example:
+In addition it is possible to append control parameters, giving the
+flag name first, followed by the parameter, all separated by a
+space. Example:
 
 ```
 #
@@ -57,14 +75,35 @@ $ ./trac_fortran data/dirlist trac.ctl atm_split.tab meteo/ei ATM_BASENAME atm_d
 
 ### mptrac_fortran.f90
 
[…]
