### 2025-04-22 3836ad1
Update mkdocs manual.
