+++ docs/manual/profiling.md
@@ -2,17 +2,30 @@
 
 ## Introduction
 
-In software engineering, profiling ("program profiling", "software profiling") is a form of dynamic program analysis that measures, for example, the space (memory) or time complexity of a program, the usage of particular instructions, or the frequency and duration of function calls. Most commonly, profiling information serves to aid program optimization.
+In software engineering, *profiling* is a form of dynamic program
+analysis that measures, for example, the space (memory) or time
+complexity of a program, the usage of particular instructions, or the
+frequency and duration of function calls. Most commonly, profiling
+information serves to aid program optimization.
 
-Profiling is achieved by instrumenting either the program source code or its binary executable form using a tool called a profiler (or code profiler). Profilers may use a number of different techniques, such as event-based, statistical, instrumented, and simulation methods.
+Profiling is achieved by instrumenting either the program's source
+code or its binary executable with a tool called a profiler (or code
+profiler). Profilers may use a number of different techniques, such as
+event-based, statistical, instrumentation-based, and simulation
+methods.
 
-Profilers use a wide variety of techniques to collect data, including hardware interrupts, code instrumentation, instruction set simulation, operating system hooks, and performance counters. Profilers are used in the performance engineering process.
+Profilers use a wide variety of techniques to collect data, including
+hardware interrupts, code instrumentation, instruction set simulation,
+operating system hooks, and performance counters. Profilers are
+commonly used in the performance engineering process.
 
-## Timers of the trac tool
+## Timers in MPTRAC
 
-A set of timers has been implemented directly in the trac tool. Next to the runtime for different parts of the code, also estimates of the memory needs and the problem size are reported.
+A set of timers has been directly implemented in the `trac` tool. In
+addition to the runtime for different parts of the code, estimates of
+memory usage and problem size are also reported.
 
-Example of output reported by the trac tool:
+An example of output reported by the `trac` tool:
 
 ```
 SIZE_NP = 10000
@@ -75,21 +88,32 @@ TIMER_GROUP_OUTPUT = 8.171 s
 TIMER_TOTAL = 37.936 s
 ```
 
-SIZE_NP refers to the number of particles used in the simulation. SIZE_MPI_TASKS refers to the number of MPI tasks. SIZE_OMP_THREADS refers to the number of OpenMP threads per MPI task. SIZE_ACC_DEVICES refers to the number of GPU devices per MPI task.
+`SIZE_NP` refers to the number of particles used in the
+simulation. `SIZE_MPI_TASKS` refers to the number of MPI
+tasks. `SIZE_OMP_THREADS` refers to the number of OpenMP threads per
+MPI task. `SIZE_ACC_DEVICES` refers to the number of GPU devices per
+MPI task.
 
-The memory needs have been derived by analyzing the data structures used in the code and are not based on measurements.
+The memory requirements have been derived by analyzing the data
+structures used in the code and are not based on measurements.
 
-Some timers are zero or not shown because the corresponding modules have not been used in this run.
+Some timers are zero or not shown because the corresponding modules
+have not been used in this run.
 
-A simple optimization of memory needs (and runtime!) is to adjust the constants EX, EY, and EZ defined in mptrac.h to match the grid dimensions of the meteorological data set and the constant NP to match the number of particles used in the simulation.
+A simple optimization of memory requirements (and runtime!) is to
+adjust the constants `EX`, `EY`, and `EP`, defined in `mptrac.h`, to
+match the grid dimensions of the meteorological data set and the
+constant `NP` to match the number of particles used in the simulation.
 
 ## Profiling of CPU runs
 
[…]
