### 2025-04-25 #8031 @AnjaliSandip
👋@dostuffthatmatters, @dorchard, @platipodium, @michaeldenes, and @EmilyBourne would you review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines:[ https://joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html).
