+++ projects/web_runner/requirements.txt
@@ -1,0 +1,3 @@
+flask
+matplotlib
+cartopy
