+++ projects/web_runner/app.py
@@ -1,0 +1,310 @@
+from flask import Flask, render_template, request, send_from_directory, jsonify
+import subprocess
+import threading
+import os
+import uuid
+import shutil
+from datetime import datetime, timezone
+import io
+import base64
+import glob
+import numpy as np
+import matplotlib
+matplotlib.use('Agg')
+import matplotlib.pyplot as plt
+import cartopy.crs as ccrs
+import cartopy.feature as cfeature
+
+# Maximum number of simulations running at once:
+process_semaphore = threading.Semaphore(3)
+
+# Flask app:
+app = Flask(__name__)
+
+
+def run_command(cmd, timeout=300):
+    """
+    Runs a command with concurrency limit and timeout.
+    Combines stdout and stderr into a single output stream.
+    """
+    acquired = process_semaphore.acquire(timeout=5)
+    if not acquired:
+        return -999, "❌ Server is too busy."
+
+    try:
+        result = subprocess.run(
+            cmd,
+            stdout=subprocess.PIPE,
+            stderr=subprocess.STDOUT,
+            text=True,
+            timeout=timeout
+        )
+        return result.returncode, result.stdout
+    except subprocess.TimeoutExpired as e:
+        return -1, (e.stdout or '') + "\n❌ Command timed out after {} seconds.".format(timeout)
+    finally:
+        process_semaphore.release()
+
+        
+def seconds_since_2000(time_str):
+    """
+    Converts a UTC time string 'YYYY-MM-DD HH:MM:SS'
+    to seconds since January 1, 2000, 00:00 UTC.
+    """
+    dt = datetime.strptime(time_str, "%Y-%m-%d %H:%M:%S")
+    dt = dt.replace(tzinfo=timezone.utc)
+    ref = datetime(2000, 1, 1, 0, 0, 0, tzinfo=timezone.utc)
+    delta = dt - ref
+    return delta.total_seconds()
+
+
+def create_plot(lons, lats, heights):
+    """
+    Create a scatter plot with longitude and latitude, color-coded by height.
+    Returns the plot as a base64 string to be embedded in an HTML page.
+    """
+
+    fig = plt.figure(figsize=(15, 12))
+    ax = plt.axes(projection=ccrs.PlateCarree())
+    ax.set_global()
+
+    # Scatter plot
+    sc = ax.scatter(
+        lons, lats, c=heights, cmap='viridis',
+        s=10, edgecolors='none', transform=ccrs.PlateCarree()
+    )
+
+    # Add coastlines
+    ax.coastlines(color='white')
+
+    # Add a grid
+    gl = ax.gridlines(draw_labels=True, color='gray', linestyle='--', alpha=0.5)
+    
+    # Customize gridline labels
+    gl.xlabel_style = {'color': 'white'}
+    gl.ylabel_style = {'color': 'white'}
+
+    # Set background colors
+    ax.set_facecolor('#222')
+    fig.patch.set_facecolor('#111')
+
+    # Colorbar
+    cbar = plt.colorbar(sc, orientation='horizontal', pad=0.05, aspect=50)
+    cbar.set_label('Height (km)', color='white')
+    cbar.ax.xaxis.set_tick_params(color='white')
+    plt.setp(cbar.ax.get_xticklabels(), color='white')
+
+    # Set axis tick colors
+    ax.xaxis.label.set_color('white')
+    ax.yaxis.label.set_color('white')
+    ax.tick_params(axis='both', colors='white')
+
+    # Title
+    ax.set_title('MPTRAC | Air Parcel Trajectories', fontsize=16, color='white')
+
+    # Save to base64
+    img = io.BytesIO()
[…]
