+++ projects/web_runner/app.py
@@ -4,6 +4,7 @@ import threading
 import os
 import uuid
 import shutil
+import time
 from datetime import datetime, timezone
 import io
 import base64
@@ -14,6 +15,17 @@ matplotlib.use('Agg')
 import matplotlib.pyplot as plt
 import cartopy.crs as ccrs
 import cartopy.feature as cfeature
+import textwrap
+
+# === Configuration ===
+ATM_INIT_CMD = os.environ.get('ATM_INIT_CMD', '/home/lars/wrk/mptrac/src/atm_init')
+TRAC_CMD = os.environ.get('TRAC_CMD', '/home/lars/wrk/mptrac/src/trac')
+METBASE = os.environ.get('METBASE', '/home/lars/wrk/mptrac/tests/data/ei')
+
+RUNS_DIR = 'runs/working'
+ZIPS_DIR = 'runs/zips'
+os.makedirs(RUNS_DIR, exist_ok=True)
+os.makedirs(ZIPS_DIR, exist_ok=True)
 
 # Maximum number of simulations running at once:
 process_semaphore = threading.Semaphore(3)
@@ -22,11 +34,16 @@ process_semaphore = threading.Semaphore(3)
 app = Flask(__name__)
 
 
+def delayed_cleanup(path, delay=600):
+    def cleanup():
+        time.sleep(delay)
+        if os.path.exists(path):
+            shutil.rmtree(path)
+    threading.Thread(target=cleanup, daemon=True).start()
+
+
 def run_command(cmd, timeout=300):
-    """
-    Runs a command with concurrency limit and timeout.
-    Combines stdout and stderr into a single output stream.
-    """
+    """Runs a command with concurrency limit and timeout."""
     acquired = process_semaphore.acquire(timeout=5)
     if not acquired:
         return -999, "❌ Server is too busy."
@@ -41,68 +58,45 @@ def run_command(cmd, timeout=300):
         )
         return result.returncode, result.stdout
     except subprocess.TimeoutExpired as e:
-        return -1, (e.stdout or '') + "\n❌ Command timed out after {} seconds.".format(timeout)
+        return -1, (e.stdout or '') + f"\n❌ Command timed out after {timeout} seconds."
     finally:
         process_semaphore.release()
 
-        
+
 def seconds_since_2000(time_str):
-    """
-    Converts a UTC time string 'YYYY-MM-DD HH:MM:SS'
-    to seconds since January 1, 2000, 00:00 UTC.
-    """
+    """Converts a UTC time string 'YYYY-MM-DD HH:MM:SS' to seconds since January 1, 2000."""
     dt = datetime.strptime(time_str, "%Y-%m-%d %H:%M:%S")
     dt = dt.replace(tzinfo=timezone.utc)
     ref = datetime(2000, 1, 1, 0, 0, 0, tzinfo=timezone.utc)
-    delta = dt - ref
-    return delta.total_seconds()
+    return (dt - ref).total_seconds()
 
 
 def create_plot(lons, lats, heights):
-    """
-    Create a scatter plot with longitude and latitude, color-coded by height.
-    Returns the plot as a base64 string to be embedded in an HTML page.
-    """
-
+    """Create a scatter plot."""
     fig = plt.figure(figsize=(15, 12))
     ax = plt.axes(projection=ccrs.PlateCarree())
     ax.set_global()
 
-    # Scatter plot
-    sc = ax.scatter(
-        lons, lats, c=heights, cmap='viridis',
-        s=10, edgecolors='none', transform=ccrs.PlateCarree()
-    )
+    sc = ax.scatter(lons, lats, c=heights, cmap='viridis', s=10, edgecolors='none', transform=ccrs.PlateCarree())
 
-    # Add coastlines
     ax.coastlines(color='white')
-
-    # Add a grid
[…]
