### 2025-04-29 369560a
Added link to AI documentation.
