### 2025-04-29 529fa80
Fix for plotting a single trajectory.
