### 2025-04-29 ea7887c
Added check of total number of trajectories.
