### 2025-04-30 a77b989
Made app.py more concise.
