+++ projects/web_runner/app.py
@@ -1,195 +1,103 @@
-from flask import Flask, render_template, request, send_from_directory, jsonify
-import subprocess
-import threading
-import os
-import uuid
-import shutil
-import time
-from datetime import datetime, timezone
-import io
-import base64
-import glob
+from flask import Flask, render_template, request, send_from_directory
+import subprocess, threading, os, uuid, shutil, time, io, base64, glob, textwrap
 import numpy as np
+from datetime import datetime, timezone
 import matplotlib
 matplotlib.use('Agg')
 import matplotlib.pyplot as plt
 import cartopy.crs as ccrs
 import cartopy.feature as cfeature
-import textwrap
 
-# === Configuration ===
-ATM_INIT_CMD = os.environ.get('ATM_INIT_CMD', '/home/lars/wrk/mptrac/src/atm_init')
-TRAC_CMD = os.environ.get('TRAC_CMD', '/home/lars/wrk/mptrac/src/trac')
-METBASE = os.environ.get('METBASE', '/home/lars/wrk/mptrac/tests/data/ei')
-
-RUNS_DIR = 'runs/working'
-ZIPS_DIR = 'runs/zips'
+# === Config ===
+ATM_INIT_CMD = os.getenv('ATM_INIT_CMD', '/home/lars/wrk/mptrac/src/atm_init')
+TRAC_CMD = os.getenv('TRAC_CMD', '/home/lars/wrk/mptrac/src/trac')
+METBASE = os.getenv('METBASE', '/home/lars/wrk/mptrac/tests/data/ei')
+RUNS_DIR, ZIPS_DIR = 'runs/working', 'runs/zips'
 os.makedirs(RUNS_DIR, exist_ok=True)
 os.makedirs(ZIPS_DIR, exist_ok=True)
-
-# Maximum number of simulations running at once:
 process_semaphore = threading.Semaphore(3)
-
-# Flask app:
 app = Flask(__name__)
 
-
 def delayed_cleanup(path, delay=600):
-    def cleanup():
-        time.sleep(delay)
-        if os.path.exists(path):
-            shutil.rmtree(path)
-    threading.Thread(target=cleanup, daemon=True).start()
-
+    threading.Thread(target=lambda: (time.sleep(delay), shutil.rmtree(path, ignore_errors=True)), daemon=True).start()
 
 def run_command(cmd, timeout=300):
-    """Runs a command with concurrency limit and timeout."""
-    acquired = process_semaphore.acquire(timeout=5)
-    if not acquired:
+    if not process_semaphore.acquire(timeout=5):
         return -999, "❌ Server is too busy."
-
     try:
-        result = subprocess.run(
-            cmd,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.STDOUT,
-            text=True,
-            timeout=timeout
-        )
+        result = subprocess.run(cmd, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, text=True, timeout=timeout)
         return result.returncode, result.stdout
     except subprocess.TimeoutExpired as e:
         return -1, (e.stdout or '') + f"\n❌ Command timed out after {timeout} seconds."
     finally:
         process_semaphore.release()
 
-
 def seconds_since_2000(time_str):
-    """Converts a UTC time string 'YYYY-MM-DD HH:MM:SS' to seconds since January 1, 2000."""
-    dt = datetime.strptime(time_str, "%Y-%m-%d %H:%M:%S")
-    dt = dt.replace(tzinfo=timezone.utc)
-    ref = datetime(2000, 1, 1, 0, 0, 0, tzinfo=timezone.utc)
-    return (dt - ref).total_seconds()
-
[…]
