### 2025-05-02 2d1f21d
Added control parameter MET_ZSTD_LEVEL to set compression level. Re-enabled multi-threading.
