### 2025-05-02 75ab763
Added control parameter MET_ZSTD_LEVEL to set compression level. Re-enabled multi-threading.
