+++ src/mptrac.c
@@ -848,6 +848,7 @@ void compress_zstd(
   float *array,
   const size_t n,
   const int decompress,
+  const int level,
   FILE *inout) {
 
   /* Get buffer sizes... */
@@ -858,6 +859,12 @@ void compress_zstd(
   /* Allocate... */
   char *compr = (char *) calloc((uint) comprLen, 1);
   char *uncompr = (char *) array;
+  ZSTD_CCtx *cctx = ZSTD_createCCtx();
+
+  /* Enable threading and set compression level... */
+  int threads = MAX(1, omp_get_max_threads());
+  ZSTD_CCtx_setParameter(cctx, ZSTD_c_nbWorkers, threads);
+  ZSTD_CCtx_setParameter(cctx, ZSTD_c_compressionLevel, level);
 
   /* Read compressed stream and decompress array... */
   if (decompress) {
@@ -876,7 +883,7 @@ void compress_zstd(
 
   /* Compress array and output compressed stream... */
   else {
-    compsize = ZSTD_compress(compr, comprLen, uncompr, uncomprLen, 0);
+    compsize = ZSTD_compress2(cctx, compr, comprLen, uncompr, uncomprLen);
     if (ZSTD_isError(compsize)) {
       ERRMSG("Compression failed!");
     } else {
@@ -892,6 +899,7 @@ void compress_zstd(
 
   /* Free... */
   free(compr);
+  free(cctx);
 }
 #endif
 
@@ -4799,6 +4807,8 @@ void mptrac_read_ctl(
     (int) scan_ctl(filename, argc, argv, "MET_NC_LEVEL", -1, "0", NULL);
   ctl->met_nc_quant =
     (int) scan_ctl(filename, argc, argv, "MET_NC_QUANT", -1, "0", NULL);
+  ctl->met_zstd_level =
+    (int) scan_ctl(filename, argc, argv, "MET_ZSTD_LEVEL", -1, "0", NULL);
   ctl->met_zfp_prec =
     (int) scan_ctl(filename, argc, argv, "MET_ZFP_PREC", -1, "8", NULL);
   ctl->met_zfp_tol_t =
@@ -6637,7 +6647,7 @@ void read_met_bin_3d(
   else if (ctl->met_type == 4) {
 #ifdef ZSTD
     compress_zstd(varname, help, (size_t) (met->np * met->ny * met->nx), 1,
-		  in);
+		  ctl->met_zstd_level, in);
 #else
     ERRMSG("MPTRAC was compiled without zstd compression!");
 #endif
@@ -10653,7 +10663,7 @@ void write_met_bin_3d(
 #ifdef ZSTD
   else if (ctl->met_type == 4)
     compress_zstd(varname, help, (size_t) (met->np * met->ny * met->nx), 0,
-		  out);
+		  ctl->met_zstd_level, out);
 #endif
 
   /* Write cmultiscale data... */
