### 2025-05-02 35db274
Went back to using ZSTD_compress().
