### 2025-05-02 ae6f5f7
Went back to using ZSTD_compress().
