+++ src/mptrac.c
@@ -859,12 +859,6 @@ void compress_zstd(
   /* Allocate... */
   char *compr = (char *) calloc((uint) comprLen, 1);
   char *uncompr = (char *) array;
-  ZSTD_CCtx *cctx = ZSTD_createCCtx();
-
-  /* Enable threading and set compression level... */
-  int threads = MIN(1, omp_get_max_threads());
-  ZSTD_CCtx_setParameter(cctx, ZSTD_c_nbWorkers, threads);
-  ZSTD_CCtx_setParameter(cctx, ZSTD_c_compressionLevel, level);
 
   /* Read compressed stream and decompress array... */
   if (decompress) {
@@ -883,7 +877,7 @@ void compress_zstd(
 
   /* Compress array and output compressed stream... */
   else {
-    compsize = ZSTD_compress2(cctx, compr, comprLen, uncompr, uncomprLen);
+    compsize = ZSTD_compress(compr, comprLen, uncompr, uncomprLen, level);
     if (ZSTD_isError(compsize)) {
       ERRMSG("Compression failed!");
     } else {
@@ -899,7 +893,6 @@ void compress_zstd(
 
   /* Free... */
   free(compr);
-  free(cctx);
 }
 #endif
 
