### 2025-05-04 c089023
Added alternative variable names for hybrid level coefficients.
