### 2025-05-04 06f271e
Added alternative variable names for hybrid level coefficients.
