### 2025-05-06 #8031 @KYANJO
That would work for me

Brian KYANJO, PhD
Postdoctoral fellow, EAS Georgia Tech
GitHub<https://github.com/KYANJO> | Orcid<https://orcid.org/0000-0002-0995-1051>
