### 2025-05-07 #8177 @editorialbot
[REVIEW]: MPTRAC: A high-performance Lagrangian transport model for atmospheric air parcel dispersion
**Submitting author:** @lars2015 (<a href="http://orcid.org/0000-0003-3773-4377">Lars Hoffmann</a>)
**Repository:** https://github.com/slcs-jsc/mptrac
**Branch with paper.md** (empty if default branch):
**Version:** v3.0
