### 2025-05-15 slcs-jsc/mptrac#49 @dostuffthatmatters
[JOSS REVIEW]: Transport Model Validation
*This issue is part of the JOSS review of MPTRAC (https://github.com/openjournals/joss-reviews/issues/8177)*

Dear @lars2015,

first of all, the whole model and its documentation is very impressive. The documentation is on point and the usability is great. Most of the issues I will add soon are only about adding a bit of information to the documentation or the manuscript. Great work!

Many LPDMs are validated using controlled release experiments. Like the following study evaluating HYSPLIT, STILT, and Flexpart with observational data from CAPTEX and ANATEX: https://doi.org/10.1175/JAMC-D-13-0125.1 This is definitely not required for the JOSS publication but I was wondering if anyone has ever done that for MPTrac. The Model physics looks to be on par or even more advanced than the physics of other models, but an actual validation could increase the adoption of MPTrac.

I think, mentioning model validation strategies or a possible future validation in the documentation would be very beneficial.

Best,
Moritz Makowski

Doctoral Candidate
Environmental Sensing and Modeling (Prof. Jia Chen)
Technical University of Munich
