### 2025-05-15 slcs-jsc/mptrac#50 @dostuffthatmatters
[JOSS REVIEW]: Parquet Trajectory Outputs
*This issue is part of the JOSS review of auto-icon (https://github.com/openjournals/joss-reviews/issues/8177)*

Would it be possible to write particle trajectories as parquet files?

We sometimes encounter performance problems with some NetCDF library distributions. But writing everything out as ASCII or binary will use much more storage than required – our compressed trajectories are only 10-20% of the file size as the ASCII ones. Parquet has the benefit of being a well adopted standard for tabular data like CSV but also ensuring column-type safety and compression.
