+++ doxygen/search/variables_f.js
@@ -1,0 +1,6 @@
+var searchData=
+[
+  ['rng_5ftype_0',['rng_type',['../structctl__t.html#a42c9413e93cc78f0e6ad427693a4e796',1,'ctl_t']]],
+  ['rs_1',['rs',['../structcache__t.html#aa35b7fdf9768493f676a4e24ad78e4d9',1,'cache_t']]],
+  ['rwc_2',['rwc',['../structmet__t.html#a05425967991c53f965f2551eb12321d8',1,'met_t']]]
+];
