### 2025-05-23 #8177 @AnjaliSandip
@filippi , your review will be checklist based. You will have a separate checklist that you should update when carrying out your review. First of all you need to run this command in a separate comment to create the checklist:

The reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. Any questions/concerns please let me know.

✨ Please start on your review when you are able, and be sure to complete your review in the next six weeks, at the very latest ✨
