### 2025-05-23 slcs-jsc/mptrac#51 @dostuffthatmatters
[JOSS REVIEW]: Model Limitations Within PBL
*This issue is part of the JOSS review of MPTRAC (https://github.com/openjournals/joss-reviews/issues/8177)*

In the paper you show an example of particles being released at the surface and convected into the free troposphere. Yet, in the documentation and the README you mention that MPTRAC is made "to analyze atmospheric transport processes in the free troposphere and stratosphere."

From the model physics I could not tell, why the model is not able to model transport within the PBL. Could you elaborate on that? It is totally valid not to put a focus on PBL transport, but it would be important to know the limitations of the model from the paper and/or its documenation.
