### 2025-05-30 slcs-jsc/mptrac#52 @dostuffthatmatters
[JOSS REVIEW]: Clarification on pressure level interpolation
*This issue is part of the JOSS review of MPTRAC (https://github.com/openjournals/joss-reviews/issues/8177)*

In the documentation, it states:

> It is possible to provide meteo data on model levels rather than pressure levels. In this case, pressure on model levels (PL) is required as an additional 3-D input variable in order to perform a vertical interpolation from model levels to pressure levels.

The model still runs on pressure levels, right? How do you determine the pressure levels to be used? If they are significantly coarser than the met data, you might have very strong interpolation artifacts.

Is it technically possible (in the future) to run the model on model levels? Because that has been the strategy with many LPDM runs using partially turbulence resolving meteorological models.

**EDIT:** I overlooked the example of ECMWF to pressure-level interpolation. So I assume the model just uses the pressure levels you give it. The second question remains.
