### 2025-06-05 26989ea
Update write_csi_ens() for esemble processing of some scores
