+++ src/mptrac.c
@@ -3381,40 +3381,82 @@ void module_mixing(
   }
 
   /* Calculate interparcel mixing... */
-  if (ctl->qnt_m >= 0)
-    module_mixing_help(ctl, clim, atm, ixs, iys, izs, ctl->qnt_m);
-  if (ctl->qnt_vmr >= 0)
-    module_mixing_help(ctl, clim, atm, ixs, iys, izs, ctl->qnt_vmr);
-  if (ctl->qnt_Ch2o >= 0)
-    module_mixing_help(ctl, clim, atm, ixs, iys, izs, ctl->qnt_Ch2o);
-  if (ctl->qnt_Co3 >= 0)
-    module_mixing_help(ctl, clim, atm, ixs, iys, izs, ctl->qnt_Co3);
-  if (ctl->qnt_Cco >= 0)
-    module_mixing_help(ctl, clim, atm, ixs, iys, izs, ctl->qnt_Cco);
-  if (ctl->qnt_Coh >= 0)
-    module_mixing_help(ctl, clim, atm, ixs, iys, izs, ctl->qnt_Coh);
-  if (ctl->qnt_Ch >= 0)
-    module_mixing_help(ctl, clim, atm, ixs, iys, izs, ctl->qnt_Ch);
-  if (ctl->qnt_Cho2 >= 0)
-    module_mixing_help(ctl, clim, atm, ixs, iys, izs, ctl->qnt_Cho2);
-  if (ctl->qnt_Ch2o2 >= 0)
-    module_mixing_help(ctl, clim, atm, ixs, iys, izs, ctl->qnt_Ch2o2);
-  if (ctl->qnt_Co1d >= 0)
-    module_mixing_help(ctl, clim, atm, ixs, iys, izs, ctl->qnt_Co1d);
-  if (ctl->qnt_Co3p >= 0)
-    module_mixing_help(ctl, clim, atm, ixs, iys, izs, ctl->qnt_Co3p);
-  if (ctl->qnt_Cccl4 >= 0)
-    module_mixing_help(ctl, clim, atm, ixs, iys, izs, ctl->qnt_Cccl4);
-  if (ctl->qnt_Cccl3f >= 0)
-    module_mixing_help(ctl, clim, atm, ixs, iys, izs, ctl->qnt_Cccl3f);
-  if (ctl->qnt_Cccl2f2 >= 0)
-    module_mixing_help(ctl, clim, atm, ixs, iys, izs, ctl->qnt_Cccl2f2);
-  if (ctl->qnt_Cn2o >= 0)
-    module_mixing_help(ctl, clim, atm, ixs, iys, izs, ctl->qnt_Cn2o);
-  if (ctl->qnt_Csf6 >= 0)
-    module_mixing_help(ctl, clim, atm, ixs, iys, izs, ctl->qnt_Csf6);
-  if (ctl->qnt_aoa >= 0)
-    module_mixing_help(ctl, clim, atm, ixs, iys, izs, ctl->qnt_aoa);
+  if (ctl->nens > 0)
+  {
+    if (ctl->qnt_m >= 0)
+      module_ens_mixing_help(ctl, clim, atm, ixs, iys, izs, ctl->qnt_m);
+    if (ctl->qnt_vmr >= 0)
+      module_ens_mixing_help(ctl, clim, atm, ixs, iys, izs, ctl->qnt_vmr);
+    if (ctl->qnt_Ch2o >= 0)
+      module_ens_mixing_help(ctl, clim, atm, ixs, iys, izs, ctl->qnt_Ch2o);
+    if (ctl->qnt_Co3 >= 0)
[…]
