### 2025-06-05 c8d1cf5
Update write_csi_ens() for esemble processing of some scores
