### 2025-06-07 10a83d8
Fixed GPU code for ensemble output.
