### 2025-06-07 970c9c3
Fixed GPU code for ensemble output.
