+++ src/mptrac.c
@@ -2018,12 +2018,12 @@ void level_definitions(
       ctl->met_p[ctl->met_np - ip - 1] = press[ip];
 
   } else {
-    ERRMSG("Use 0 for l137, 1 for l91, 2 for l60 or values between 3 and 7.")
+    ERRMSG("Use 0 for l137, 1 for l91, 2 for l60 or values between 3 and 7.");
   }
 
-  if (ctl->met_np > NP)
-    ERRMSG("Recompile with larger NP to use this pressure level definition.")
-	    
+  if (ctl->met_np > NP) {
+    ERRMSG("Recompile with larger NP to use this pressure level definition.");
+  }
 }
 
 /*****************************************************************************/
@@ -2275,8 +2275,9 @@ void module_advect(
 
       /* Convert zeta to pressure... */
       intpol_met_4d_coord(met0, met0->zetal, met0->pl, met1, met1->zetal,
-			  met1->pl, atm->time[ip], atm->q[ctl->qnt_zeta][ip],
-			  atm->lon[ip], atm->lat[ip], &atm->p[ip], ci, cw, 1);
+			  met1->pl, atm->time[ip],
+			  atm->q[ctl->qnt_zeta][ip], atm->lon[ip],
+			  atm->lat[ip], &atm->p[ip], ci, cw, 1);
     }
   }
 }
@@ -2652,8 +2653,8 @@ void module_convection(
       double tbot, ttop;
       intpol_met_time_3d(met0, met0->t, met1, met1->t, atm->time[ip],
 			 pbot, atm->lon[ip], atm->lat[ip], &tbot, ci, cw, 1);
-      intpol_met_time_3d(met0, met0->t, met1, met1->t, atm->time[ip],
-			 ptop, atm->lon[ip], atm->lat[ip], &ttop, ci, cw, 1);
+      intpol_met_time_3d(met0, met0->t, met1, met1->t, atm->time[ip], ptop,
+			 atm->lon[ip], atm->lat[ip], &ttop, ci, cw, 1);
       const double rhobot = pbot / tbot;
       const double rhotop = ptop / ttop;
 
@@ -3302,9 +3303,9 @@ void module_meteo(
   PARTICLE_LOOP(0, atm->np, 0,
 		"acc data present(ctl,cache,clim,met0,met1,atm)") {
 
-    double ps, ts, zs, us, vs, ess, nss, shf, lsm, sst, pbl, pt, pct, pcb, cl,
-      plcl, plfc, pel, cape, cin, o3c, pv, t, tt, u, v, w, h2o, h2ot, o3, lwc,
-      rwc, iwc, swc, cc, z, zt;
+    double ps, ts, zs, us, vs, ess, nss, shf, lsm, sst, pbl, pt, pct, pcb,
+      cl, plcl, plfc, pel, cape, cin, o3c, pv, t, tt, u, v, w, h2o, h2ot,
+      o3, lwc, rwc, iwc, swc, cc, z, zt;
 
     /* Interpolate meteo data... */
     INTPOL_INIT;
@@ -3430,8 +3431,7 @@ void module_mixing(
   }
 
   /* Calculate interparcel mixing... */
-  if (ctl->nens > 0)
-  {
+  if (ctl->nens > 0) {
     if (ctl->qnt_m >= 0)
       module_ens_mixing_help(ctl, clim, atm, ixs, iys, izs, ctl->qnt_m);
     if (ctl->qnt_vmr >= 0)
@@ -3465,10 +3465,8 @@ void module_mixing(
     if (ctl->qnt_Csf6 >= 0)
       module_ens_mixing_help(ctl, clim, atm, ixs, iys, izs, ctl->qnt_Csf6);
     if (ctl->qnt_aoa >= 0)
[…]
