+++ src/mptrac.h
@@ -5017,7 +5017,43 @@ void module_chem_grid(
   atm_t * atm,
   const double t);
 
-void module_ens_chem_grid(
+/**
+ * @brief Processes atmospheric ensemble chemical data on a defined 3D grid.
+ *
+ * This function aggregates and interpolates mass and mixing ratio data for 
+ * atmospheric parcels over a 3D spatial grid, for multiple ensemble members.
+ * It computes grid-based statistics and updates each air parcel with the 
+ * appropriate chemical quantities (e.g., mass and concentration).
+ *
+ * @param ctl  Pointer to the model control structure containing grid and configuration parameters.
+ * @param met0 Pointer to the meteorological dataset at the beginning of the interpolation time window.
+ * @param met1 Pointer to the meteorological dataset at the end of the interpolation time window.
+ * @param atm  Pointer to the structure containing atmospheric parcel data, including positions and properties.
+ * @param tt   Simulation time (in seconds) for which the chemical grid is evaluated.
+ *
+ * @details
+ * The function performs the following main tasks:
+ * - Checks configuration and model parameters for validity.
+ * - Allocates and initializes memory for spatial grid properties (altitude, pressure, area, etc.).
+ * - Maps air parcel positions to 3D grid indices (longitude, latitude, height).
+ * - Aggregates mass into grid cells for each ensemble.
+ * - Interpolates meteorological data (e.g., temperature) at each parcel's position.
+ * - Computes volume mixing ratios using the interpolated data and grid cell properties.
+ * - Updates each air parcel with the computed chemical quantity.
+ *
+ * Parallelization is supported via OpenMP or OpenACC (depending on compilation).
+ *
+ * @note
+ * - Memory allocation is done dynamically within the function and freed before exit.
+ * - The function uses macros and helper functions like `P()`, `Z()`, `RHO()`, `ARRAY_3D()`,
+ *   and `intpol_met_time_3d()` for calculations and interpolation.
+ * - Requires ensemble quantity indices (`qnt_ens`, `qnt_m`, and `qnt_Cx`) to be set in `ctl`.
+ *
+ * @warning If molar mass is not set (`molmass <= 0`), the function will issue an error and exit.
+ *
+ * @authors Mingzhao Liu
+ */
+void module_chem_grid_ens(
   const ctl_t * ctl,
   met_t * met0,
   met_t * met1,
@@ -5497,7 +5533,41 @@ void module_mixing_help(
   const int *izs,
   const int qnt_idx);
 
-void module_ens_mixing_help(
+/**
+ * @brief Applies ensemble-based interparcel mixing for a given tracer quantity.
+ *
+ * This function calculates local grid-cell averages of a specified tracer quantity
+ * (`qnt_idx`) across ensemble members, and then applies a mixing adjustment to each 
+ * atmospheric parcel based on the difference between its value and the mean within its grid cell.
+ *
+ * @param ctl      Pointer to the model control structure, containing mixing configuration and grid dimensions.
+ * @param clim     Pointer to climatological data used to determine tropospheric weighting, if applicable.
+ * @param atm      Pointer to the structure containing atmospheric parcel data, including tracer fields.
+ * @param ixs      Array of X-indices mapping parcels to grid cells in the horizontal (longitude) direction.
+ * @param iys      Array of Y-indices mapping parcels to grid cells in the horizontal (latitude) direction.
+ * @param izs      Array of Z-indices mapping parcels to vertical levels; negative values indicate invalid/masked parcels.
+ * @param qnt_idx  Index of the tracer quantity in the `atm->q` array to be mixed.
+ *
+ * @details
+ * The function performs the following steps:
[…]
