+++ src/mptrac.c
@@ -2550,6 +2550,157 @@ void module_chem_grid(
 
 /*****************************************************************************/
 
+void module_chem_grid_ens(
+  const ctl_t *ctl,
+  met_t *met0,
+  met_t *met1,
+  atm_t *atm,
+  const double tt) {
+
+  /* Check quantities... */
+  if (ctl->qnt_m < 0 || ctl->qnt_Cx < 0)
+    return;
+  if (ctl->molmass <= 0)
+    ERRMSG("Molar mass is not defined!");
+
+  /* Set timer... */
+  SELECT_TIMER("MODULE_CHEM_GRID", "PHYSICS", NVTX_GPU);
+
+  /* Allocate... */
+  const int np = atm->np;
+  const int nz = ctl->chemgrid_nz;
+  const int nx = ctl->chemgrid_nx;
+  const int ny = ctl->chemgrid_ny;
+  const int ngrid = nx * ny * nz;
+
+  double *restrict const z = (double *) malloc((size_t) nz * sizeof(double));
+  double *restrict const press =
+    (double *) malloc((size_t) nz * sizeof(double));
+  double *restrict const mass =
+    (double *) calloc((size_t) ngrid * (size_t) ctl->nens, sizeof(double));
+  double *restrict const area =
+    (double *) malloc((size_t) ny * sizeof(double));
+  double *restrict const lon =
+    (double *) malloc((size_t) nx * sizeof(double));
+  double *restrict const lat =
+    (double *) malloc((size_t) ny * sizeof(double));
+
+  int *restrict const ixs = (int *) malloc((size_t) np * sizeof(int));
+  int *restrict const iys = (int *) malloc((size_t) np * sizeof(int));
+  int *restrict const izs = (int *) malloc((size_t) np * sizeof(int));
+
+  /* Set grid box size... */
+  const double dz = (ctl->chemgrid_z1 - ctl->chemgrid_z0) / nz;
+  const double dlon = (ctl->chemgrid_lon1 - ctl->chemgrid_lon0) / nx;
+  const double dlat = (ctl->chemgrid_lat1 - ctl->chemgrid_lat0) / ny;
+
+  /* Set vertical coordinates... */
+#ifdef _OPENACC
+#pragma acc enter data create(ixs[0:np],iys[0:np],izs[0:np],z[0:nz],press[0:nz],mass[0:ngrid],area[0:ny],lon[0:nx],lat[0:ny])
+#pragma acc data present(ctl,met0,met1,atm,ixs,iys,izs,z,press,mass,area,lon,lat)
+#pragma acc parallel loop independent gang vector
+#else
+#pragma omp parallel for default(shared)
+#endif
+  for (int iz = 0; iz < nz; iz++) {
+    z[iz] = ctl->chemgrid_z0 + dz * (iz + 0.5);
+    press[iz] = P(z[iz]);
+  }
+
+  /* Set time interval for output... */
+  const double t0 = tt - 0.5 * ctl->dt_mod;
+  const double t1 = tt + 0.5 * ctl->dt_mod;
+
+  /* Get indices... */
+#ifdef _OPENACC
+#pragma acc parallel loop independent gang vector
+#else
+#pragma omp parallel for default(shared)
+#endif
+  for (int ip = 0; ip < np; ip++) {
+    ixs[ip] = (int) ((atm->lon[ip] - ctl->chemgrid_lon0) / dlon);
[…]
