+++ doxygen/search/functions_1.js
@@ -1,0 +1,15 @@
+var searchData=
+[
+  ['cart2geo_0',['cart2geo',['../mptrac_8h.html#aaf98f7230de0893b8dc4833d3c83c7a4',1,'cart2geo(const double *x, double *z, double *lon, double *lat):&#160;mptrac.c'],['../mptrac_8c.html#aaf98f7230de0893b8dc4833d3c83c7a4',1,'cart2geo(const double *x, double *z, double *lon, double *lat):&#160;mptrac.c']]],
+  ['clim_5foh_1',['clim_oh',['../mptrac_8c.html#a9ba00889eeb600010467d8b62de94a36',1,'clim_oh(const ctl_t *ctl, const clim_t *clim, const double t, const double lon, const double lat, const double p):&#160;mptrac.c'],['../mptrac_8h.html#a9ba00889eeb600010467d8b62de94a36',1,'clim_oh(const ctl_t *ctl, const clim_t *clim, const double t, const double lon, const double lat, const double p):&#160;mptrac.c']]],
+  ['clim_5foh_5fdiurnal_5fcorrection_2',['clim_oh_diurnal_correction',['../mptrac_8c.html#aba22251f4ea0fe56c2a59719998c650a',1,'clim_oh_diurnal_correction(const ctl_t *ctl, clim_t *clim):&#160;mptrac.c'],['../mptrac_8h.html#aba22251f4ea0fe56c2a59719998c650a',1,'clim_oh_diurnal_correction(const ctl_t *ctl, clim_t *clim):&#160;mptrac.c']]],
+  ['clim_5fphoto_3',['clim_photo',['../mptrac_8c.html#a69a55a5bcd0af800ddc35cbe3b16cd15',1,'clim_photo(const double rate[CP][CSZA][CO3], const clim_photo_t *photo, const double p, const double sza, const double o3c):&#160;mptrac.c'],['../mptrac_8h.html#a69a55a5bcd0af800ddc35cbe3b16cd15',1,'clim_photo(const double rate[CP][CSZA][CO3], const clim_photo_t *photo, const double p, const double sza, const double o3c):&#160;mptrac.c']]],
+  ['clim_5ftropo_4',['clim_tropo',['../mptrac_8c.html#a4531cb279e09061a6f34a73695c7b692',1,'clim_tropo(const clim_t *clim, const double t, const double lat):&#160;mptrac.c'],['../mptrac_8h.html#a4531cb279e09061a6f34a73695c7b692',1,'clim_tropo(const clim_t *clim, const double t, const double lat):&#160;mptrac.c']]],
+  ['clim_5ftropo_5finit_5',['clim_tropo_init',['../mptrac_8c.html#aa9cf40c6960c5078d1ea9e12a2fb0872',1,'clim_tropo_init(clim_t *clim):&#160;mptrac.c'],['../mptrac_8h.html#aa9cf40c6960c5078d1ea9e12a2fb0872',1,'clim_tropo_init(clim_t *clim):&#160;mptrac.c']]],
+  ['clim_5fts_6',['clim_ts',['../mptrac_8c.html#acbf877af94269d0aa1c94b7c50121df8',1,'clim_ts(const clim_ts_t *ts, const double t):&#160;mptrac.c'],['../mptrac_8h.html#acbf877af94269d0aa1c94b7c50121df8',1,'clim_ts(const clim_ts_t *ts, const double t):&#160;mptrac.c']]],
[…]
