+++ src/mptrac.c
@@ -5604,11 +5604,14 @@ void mptrac_run_timestep(
     /* Transform from struct of array to array of struct... */
     LOG(1, "Transform from SoA to AoS.")
     atm2particles(atm, particles, *ctl);
+    //atm2particles_sort(atm, particles, *ctl, *met0); 
     
     /* Perform the communication... */
     LOG(1, "Peform the communication.")
     dd_communicate_particles( particles, &atm->np, MPI_Particle, 
       neighbours, ctl->dd_nbr_neighbours , *ctl, cache->dt);
+    //dd_communicate_particles_sorted( particles, &atm->np, MPI_Particle, 
+    //  neighbours, ctl->dd_nbr_neighbours , *ctl);      
       
     /* Transform from array of struct to struct of array... */
     LOG(1, "Transform from AoS to SoA.")
@@ -12127,10 +12130,67 @@ void write_vtk(
 
 /*****************************************************************************/
 
+void atm2particles_sort(atm_t* atm, particle_t particles[], ctl_t ctl, met_t* met) {
+
+  SELECT_TIMER("ATM2PARTICLES", "DD", NVTX_READ);
+  
+  /* Allocate... */
+  const int np = atm->np;
+  double *restrict const a = (double *) malloc((size_t) np * sizeof(double));
+  int *restrict const p = (int *) malloc((size_t) np * sizeof(int));
+  double amax = (met->nx*met->ny + met->ny)*met->np + met->np;
+  
+  /* Get the MPI rank... */
+  int rank;
+  MPI_Comm_rank(MPI_COMM_WORLD, &rank);
+  
+  /* Get flags from box index or destinations... */
+  #pragma omp parallel for default(shared)
+  for (int ip = 0; ip < np; ip++) {
+    if ((int) atm->q[ctl.qnt_destination][ip] == rank) {
+      a[ip] = (double) ((locate_reg(met->lon, met->nx, atm->lon[ip]) * met->ny +
+		 locate_irr(met->lat, met->ny, atm->lat[ip]))
+		* met->np + locate_irr(met->p, met->np, atm->p[ip]));
+    } else {
+      a[ip] = atm->q[ctl.qnt_destination][ip] + amax + 1;
+    }
+    p[ip] = ip;
+  }
+
+  /* Sorting... */
+  #ifdef THRUST
+    thrustSortWrapper(a, np, p);
+  #else
+    ERRMSG("MPTRAC was compiled without Thrust library!");
+  #endif
+  
+  /* Fill the particles buffer... */
+  //#pragma omp parallel for default(shared)
+  for (int ip = 0; ip < np; ip++) {
+
+    particles[ip].time = atm->time[p[ip]];
+    particles[ip].lon = atm->lon[p[ip]]; 
+    particles[ip].lat = atm->lat[p[ip]];
+    particles[ip].p = atm->p[p[ip]];
+
+    for (int iq = 0; iq < ctl.nq; iq++) 
+      particles[ip].q[iq] = atm->q[iq][p[ip]];
+
+  }
+  
+  /* Free... */
+  free(a);
+  free(p);
+  
+}
+
+/*****************************************************************************/
+
 void atm2particles(atm_t* atm, particle_t particles[], ctl_t ctl) {
 
   SELECT_TIMER("ATM2PARTICLES", "DD", NVTX_READ);
 
+  #pragma omp parallel for default(shared)
   for (int ip = 0; ip < atm->np; ip++) {
 
     particles[ip].time = atm->time[ip];
@@ -12151,6 +12211,7 @@ void particles2atm(atm_t* atm, particle_t particles[], ctl_t ctl) {
 
   SELECT_TIMER("PARTICLES2ATM", "DD", NVTX_READ);
   
[…]
