### 2025-06-10 slcs-jsc/mptrac#50 @lars2015
Dear Moritz, @dostuffthatmatters,

Thank you for the suggestion regarding Parquet output for trajectory data.

We agree that compressed, columnar formats like Parquet can be highly beneficial for storage efficiency and post-processing performance—particularly for large datasets. As you noted, current NetCDF and ASCII formats can be suboptimal for such use cases. Even with NetCDF, we've seen file sizes remain substantial (e.g., 4.1 GB for 10^6 particles over ten days, compared to 4.4 GB as ASCII), highlighting room for improvement.

We are currently working on enhancing NetCDF outputs by adding support for both lossless and lossy compression using the capabilities of the NetCDF library. These features are under development and should improve storage efficiency, though compression performance will depend on the use case.

Regarding Parquet specifically, we are open to exploring its adoption. One limitation is that MPTRAC is written in C, and existing Parquet libraries are primarily available for C++ or higher-level languages. Using these would require introducing C++ dependencies, which increases implementation complexity.

Nevertheless, we recognize the benefits of native support and will consider it for future releases. In the meantime, users may convert NetCDF or ASCII outputs to Parquet in post-processing.

Thank you again for the valuable suggestion.

Best regards,
Lars Hoffmann
