### 2025-06-10 slcs-jsc/mptrac#51 @lars2015
Dear Moritz, @dostuffthatmatters,

Thank you for raising this important point.

MPTRAC was originally developed with a focus on simulating transport in the free troposphere and stratosphere, where large-scale advection dominates and pressure is a suitable vertical coordinate. In its present form, MPTRAC is technically capable of computing trajectories in the planetary boundary layer (PBL), with surface pressure serving as the lower boundary of the model domain.

However, the PBL involves complex turbulence and surface-atmosphere interactions that require specialized physical parameterizations. While some relevant schemes—such as those for turbulent diffusion and convection tailored to near-surface conditions—have recently been implemented, they are still undergoing validation.

The example shown in the manuscript, involving near-surface particle releases and convective transport into the free troposphere, is included primarily for illustrative purposes. It does not yet utilize the new PBL-specific turbulence schemes and therefore should not be interpreted as a full demonstration of MPTRAC’s emerging capabilities in the PBL.

At this stage, we would like to retain the statement that MPTRAC is primarily suited for the free troposphere and stratosphere, while updating the manuscript and documentation to more clearly note that support for PBL processes is under active development.

Best regards,
Lars Hoffmann
