### 2025-06-10 slcs-jsc/mptrac#52 @lars2015
Dear Moritz, @dostuffthatmatters,

While MPTRAC traditionally runs on pressure levels, it can already perform trajectory advection using meteorological data on model levels (e.g., ECMWF's eta surfaces). This feature allows for more accurate representation of near-surface advection, helping to avoid vertical interpolation artifacts associated with coarse pressure-level data.

However, this functionality currently applies only to the advection module. Other MPTRAC components, such as diffusion and convection, still require pressure-level coordinates. Therefore, at each model timestep, vertical coordinate conversions between model and pressure levels must be performed. This setup also requires both model-level and pressure-level meteorological fields to be held in memory during simulations, and separate compute kernels for advection in each coordinate system.

While it's technically feasible to extend all modules to operate entirely on model levels (as done in models like FLEXPART), this would require significant development effort. We recognize the value of this approach and may explore it further in future versions of MPTRAC.

We will revise the manuscript and documentation accordingly to clarify the current state of model-level data support and future directions.

Best regards,
Lars Hoffmann
