### 2025-06-10 3e65c35
Fixed name of timer of module_chem_grid_ens().
