### 2025-06-11 00a1b8e
Compute pressure at half levels instead of full levels
