### 2025-06-13 2bd26da
Update JOSS paper.
