+++ docs/joss/paper.md
@@ -79,7 +79,7 @@ bibliography: paper.bib
 
 # Summary
 
-The Massive-Parallel Trajectory Calculations (MPTRAC) Lagrangian transport model [@hoffmann16; @hoffmann22] simulates the movement, dispersion, and chemical transformation of aerosol particles and trace gases in the atmosphere. Unlike grid-based models, MPTRAC tracks individual air parcels, offering high-resolution insights into transport, mixing, chemistry, and deposition. Trajectory calculations are driven by meteorological fields from data assimilation systems and forecast models. MPTRAC has been applied in various contexts, especially for tracking volcanic plumes in the troposphere and stratosphere. Optimized for high-performance computing (HPC) and graphics processing unit (GPU) systems, MPTRAC efficiently supports large-scale, high-resolution simulations, making it a powerful tool in atmospheric research.
+The Massive-Parallel Trajectory Calculations (MPTRAC) Lagrangian transport model [@hoffmann16; @hoffmann22] simulates the movement, dispersion, and chemical transformation of aerosols and trace gases in the atmosphere. Unlike grid-based models, MPTRAC tracks individual air parcels, providing high-resolution insights into transport, mixing, chemistry, and deposition. Trajectory calculations are driven by meteorological fields from data assimilation systems and forecast models. MPTRAC has been applied in various contexts, especially for tracking volcanic plumes in the troposphere and stratosphere. Optimized for high-performance computing (HPC) and graphics processing units (GPU), MPTRAC efficiently supports large-scale, high-resolution simulations, making it a powerful tool for atmospheric research.
 
 # Statement of need
 
@@ -89,7 +89,7 @@ MPTRAC supports both research and operational applications requiring high-precis
 
 # Features
 
-MPTRAC computes air parcel trajectories using wind and vertical velocity fields from global reanalysis or forecast data [@hersbach20; @gelaro17]. It accounts for eddy diffusion and subgrid-scale winds via the Langevin equation. An inter-parcel exchange module represents air mixing. Additionally, MPTRAC simulates convection, sedimentation, radioactive decay, gas and aqueous phase chemistry, and wet and dry deposition. Meteorological pre-processing routines compute boundary layer heights, CAPE, geopotential heights, and tropopause data. \autoref{fig:clusters} shows MPTRAC's geophysical and chemical modules and core software components.
+MPTRAC computes air parcel trajectories using wind and vertical velocity fields from global reanalysis or forecast data [@hersbach20; @gelaro17]. In addition to standard pressure-level input, it supports advection on forecast model levels, improving near-surface accuracy. MPTRAC accounts for eddy diffusion and subgrid-scale winds via the Langevin equation. An inter-parcel exchange module represents air mixing. MPTRAC also simulates convection, sedimentation, radioactive decay, gas and aqueous phase chemistry, and wet and dry deposition. Meteorological pre-processing routines compute boundary layer heights, CAPE, geopotential heights, and tropopause data. \autoref{fig:clusters} shows MPTRAC's geophysical and chemical modules and core software components.
 
 ![Geophysical and chemical modules and main software components of the MPTRAC Lagrangian transport model. Image adapted from @hoffmann24.\label{fig:clusters}](clusters.png){ width=95% }
 
@@ -97,15 +97,17 @@ Optimized for computational efficiency, MPTRAC features an MPI-OpenMP-OpenACC hy
 
 # Applications
 
[…]
