+++ docs/joss/paper.bib
@@ -27,6 +27,19 @@
   DOI =		 {10.5194/acp-24-763-2024}
 }
 
+@article{draxler91,
+  title =	 {Across North America tracer experiment (ANATEX):
+                  Sampling and analysis},
+  journal =	 {Atmospheric Environment. Part A. General Topics},
+  volume =	 25,
+  number =	 12,
+  pages =	 {2815-2836},
+  year =	 1991,
+  doi =		 {10.1016/0960-1686(91)90208-O},
+  author =	 {R.R. Draxler and R. Dietz and R.J. Lagomarsino and
+                  G. Start}
+}
+
 @article{gelaro17,
   author =	 "Ronald Gelaro and Will McCarty and Max J. Suárez and
                   Ricardo Todling and Andrea Molod and Lawrence Takacs
@@ -331,6 +344,21 @@
   NUMBER =	 9,
   PAGES =	 {2461--2474},
   doi =		 {10.5194/acp-5-2461-2005}
+}
+
+@article{vandop98,
+  title =	 {ETEX: A European tracer experiment; observations,
+                  dispersion modelling and emergency response},
+  journal =	 {Atmospheric Environment},
+  volume =	 32,
+  number =	 24,
+  pages =	 {4089-4094},
+  year =	 1998,
+  doi =		 {10.1016/S1352-2310(98)00248-9},
+  author =	 {H. {Van dop} and R. Addis and G. Fraser and
+                  F. Girardi and G. Graziani and Y. Inoue and N. Kelly
+                  and W. Klug and A. Kulmala and K. Nodop and
+                  J. Pretel},
 }
 
 @Article{wu17,
