### 2025-06-13 ca2bf80
Update JOSS paper.
