+++ doxygen/search/searchdata.js
@@ -1,0 +1,33 @@
+var indexSectionsWithContent =
+{
+  0: "abcdefghijklmnopqrstuvwz",
+  1: "acm",
+  2: "acdjmstw",
+  3: "bcdfgijklmnprstw",
+  4: "abcdeghiklmnopqrstuvwz",
+  5: "acdefghiklmnprstwz",
+  6: "m"
+};
+
+var indexSectionNames =
+{
+  0: "all",
+  1: "classes",
+  2: "files",
+  3: "functions",
+  4: "variables",
+  5: "defines",
+  6: "pages"
+};
+
+var indexSectionLabels =
+{
+  0: "All",
+  1: "Data Structures",
+  2: "Files",
+  3: "Functions",
+  4: "Variables",
+  5: "Macros",
+  6: "Pages"
+};
+
