### 2025-06-13 #8177 @lars2015
Dear Moritz [@dostuffthatmatters],

We've just updated the JOSS paper draft in response to your comments. We've aimed to keep the revisions concise, as we're approaching the word limit.

Please let us know if you have any further suggestions or feedback.

Thank you and best regards,
Lars Hoffmann
