### 2025-06-13 slcs-jsc/mptrac#50 @dostuffthatmatters
Thank you for the comment. I understand the technical challenges of writing Parquet from C now. NetCDF with compression also works and yes, one can do that in post-processing. I will close this issue for now since it is not that straightforward to implement.

Best,
Moritz
