### 2025-06-13 slcs-jsc/mptrac#52 @dostuffthatmatters
Great, thanks for the clarification! I think the documentation on how to do this could be a bit more clear, but the diabatic transport example with CLAMS in the documentation is enough to make it work.

Best,
Moritz
