+++ src/mptrac.c
@@ -2171,13 +2171,11 @@ void module_advect(
       /* Convert pressure to zeta... */
       INTPOL_INIT;
 
-      // TODO: can we use intpol_met_time_3d_ml instead of intpol_met_4d_coord?
-
       // Debugging:
-      //intpol_met_4d_coord(met0, met0->pl, met0->zetal, met1,
-			//  met1->pl, met1->zetal, atm->time[ip], atm->p[ip],
-			//  atm->lon[ip], atm->lat[ip],
-			//  &atm->q[ctl->qnt_zeta][ip], ci, cw, 1);
+      intpol_met_4d_coord(met0, met0->pl, met0->zetal, met1,
+			  met1->pl, met1->zetal, atm->time[ip], atm->p[ip],
+			  atm->lon[ip], atm->lat[ip],
+			  &atm->q[ctl->qnt_zeta][ip], ci, cw, 1);
 
       /* Init... */
       double dts, u[4], um = 0, v[4], vm = 0, zeta_dot[4],
@@ -3950,11 +3948,14 @@ void module_timesteps(
       cache->dt[ip] = 0.0;
 
     /* Check horizontal boundaries of local meteo data... */
-    if (ctl->dd_subdomains_meridional*ctl->dd_subdomains_zonal == 1)
+    if (ctl->dd_subdomains_meridional*ctl->dd_subdomains_zonal == 1) {
       if (local && (atm->lon[ip] <= met0->lon[0]
 		    || atm->lon[ip] >= met0->lon[met0->nx - 1]
 		    || atm->lat[ip] <= latmin || atm->lat[ip] >= latmax))
         cache->dt[ip] = 0; 
+    } else if ( (int) atm->q[ctl->qnt_subdomain][ip] == -1) {
+      cache->dt[ip] = 0;}
+      
 
   }
 }
@@ -5603,19 +5604,21 @@ void mptrac_run_timestep(
     
     /* Transform from struct of array to array of struct... */
     LOG(1, "Transform from SoA to AoS.")
-    atm2particles(atm, particles, *ctl);
-    //atm2particles_sort(atm, particles, *ctl, *met0); 
+    //atm2particles(atm, particles, *ctl)
+    int nparticles = 0;
+    atm2particles(atm, particles, *ctl, &nparticles, cache);
     
     /* Perform the communication... */
     LOG(1, "Peform the communication.")
-    dd_communicate_particles( particles, &atm->np, MPI_Particle, 
-      neighbours, ctl->dd_nbr_neighbours , *ctl, cache->dt);
-    //dd_communicate_particles_sorted( particles, &atm->np, MPI_Particle, 
-    //  neighbours, ctl->dd_nbr_neighbours , *ctl);      
+    //dd_communicate_particles( particles, &atm->np, MPI_Particle, 
+      //neighbours, ctl->dd_nbr_neighbours , *ctl, cache->dt);
+    
+    dd_communicate_particles( particles, &nparticles, MPI_Particle, 
+      neighbours, ctl->dd_nbr_neighbours , *ctl);
       
     /* Transform from array of struct to struct of array... */
     LOG(1, "Transform from AoS to SoA.")
-    particles2atm(atm, particles, *ctl); 
+    particles2atm(atm, particles, *ctl, &nparticles, cache); 
     
     LOG(1, "Free MPI datatype and particles.")
     /* Free MPI datatype... */
@@ -7473,15 +7476,15 @@ void read_met_grid(
       met->lat[0], met->lat[1], met->lat[met->ny - 1]);
 
   LOG(2, "Define subdomain properties.");
-  LOG(3, "MPI information: Rank %d, Size %d", rank, size);
[…]
