### 2025-06-17 860d72d
Some experimental GPU pragmas for the dd.
