+++ src/mptrac.c
@@ -3836,10 +3836,11 @@ void module_sort(
   const int np = atm->np;
   double *restrict const a = (double *) malloc((size_t) np * sizeof(double));
   int *restrict const p = (int *) malloc((size_t) np * sizeof(int));
+  double amax = (met0->nx*met0->ny + met0->ny)*met0->np + met0->np;
 
 #ifdef _OPENACC
-#pragma acc enter data create(a[0:np],p[0:np])
-#pragma acc data present(ctl,met0,atm,a,p)
+#pragma acc enter data create(a[0:np],p[0:np],amax)
+#pragma acc data present(ctl,met0,atm,a,p,amax)
 #endif
 
   /* Get box index... */
@@ -3849,12 +3850,17 @@ void module_sort(
 #pragma omp parallel for default(shared)
 #endif
   for (int ip = 0; ip < np; ip++) {
-    a[ip] =
-      (double) ((locate_reg(met0->lon, met0->nx, atm->lon[ip]) * met0->ny +
+    if ((int) atm->q[ctl->qnt_subdomain][ip] != -1)
+      a[ip] = 0;
+      /*(double) ((locate_reg(met0->lon, met0->nx, atm->lon[ip]) * met0->ny +
 		 locate_irr(met0->lat, met0->ny, atm->lat[ip]))
-		* met0->np + locate_irr(met0->p, met0->np, atm->p[ip]));
+		* met0->np + locate_irr(met0->p, met0->np, atm->p[ip]));*/
+    else
+      a[ip] = amax+1;
     p[ip] = ip;
   }
+  
+
 
   /* Sorting... */
 #ifdef _OPENACC
@@ -3873,10 +3879,19 @@ void module_sort(
   module_sort_help(atm->lat, p, np);
   for (int iq = 0; iq < ctl->nq; iq++)
     module_sort_help(atm->q[iq], p, np);
+    
+ /* Reset the size... */
+ int npt = 0;
+ #pragma omp parallel for reduction(+:npt)
+ for (int ip = 0; ip < np; ip++)
+   if ((int) atm->q[ctl->qnt_subdomain][ip] != -1)
+     npt++;
+     
+ atm->np = npt;
 
   /* Free... */
 #ifdef _OPENACC
-#pragma acc exit data delete(a,p)
+#pragma acc exit data delete(a,p,amax)
 #endif
   free(a);
   free(p);
@@ -5480,10 +5495,6 @@ void mptrac_run_timestep(
   /* Set time steps of air parcels... */
   module_timesteps(ctl, cache, *met0, atm, t);
 
-  /* Sort particles... */
-  if (ctl->sort_dt > 0 && fmod(t, ctl->sort_dt) == 0)
-    module_sort(ctl, *met0, atm);
-
   /* Check positions (initial)... */
   module_position(cache, *met0, *met1, atm);
 
@@ -5568,70 +5579,87 @@ void mptrac_run_timestep(
   int rankd = 0;
   MPI_Comm_rank(MPI_COMM_WORLD,&rankd);
 
-  //if ( rankd==0 ) {
-  //  printf("Before communicator: %f,%f,%f,%f \n",atm->lon[0], atm->lat[0], atm->p[0],atm->q[ctl->qnt_zeta][0]);
-  //}
-
   /* Domain decomposition... */
+  //module_dd()
   if (ctl->dd_subdomains_meridional*ctl->dd_subdomains_zonal > 1) {
-    //module_dd()
+    
     
     /* Initialize particles locally... */
+    int nparticles = 0;
+    int neighbours[NNMAX];
     particle_t* particles;
     ALLOC(particles, particle_t, NP);
-  
+
     /* Get the MPI information... */
-    int rank;
+    int rank, size;
+    
[…]
