### 2025-06-21 dbb0982
Added script to retrieve ECMWF forecasts.
