### 2025-06-21 24e22eb
Update documentation of meteo data retrieval scripts.
