+++ projects/meteo/get_merra2.sh
@@ -3,31 +3,34 @@
 # Check arguments...
 if [ $# -ne 4 ] ; then
     cat <<EOF
-MERRA-2 data download
+MERRA-2 Data Download
 
-usage: $0 <year> <month> <day> <dir>
+Usage:
+  $0 <year> <month> <day> <dir>
 
-This script retrieves instantaneous 3-dimensional, 3-hourly MERRA-2
-data (https://doi.org/10.5067/WWQSXQ8IVFW8) from NASA's GES DISC data
-archive and convert it for use with MPTRAC.
+Description:
+  This script retrieves instantaneous 3-dimensional, 3-hourly MERRA-2
+  data from NASA's GES DISC archive
+  (https://doi.org/10.5067/WWQSXQ8IVFW8) and converts it for use
+  with MPTRAC.
 
-This script retrieves the data for a given day (<year>, <month>,
-<day>) and place the converted data files in the archive directory
-(<dir>).
+  It downloads the data for the specified date (<year>, <month>, <day>)
+  and stores the converted NetCDF files in the specified output
+  directory (<dir>).
 
-Note that the script retrieves data on model levels. You will need to
-interpolate it to pressure levels using MPTRAC's internal
-interpolation functions (via the MET_NP and MET_P control parameters).
+  Note: The data are provided on model levels. You will need to
+  interpolate to pressure levels using MPTRAC's internal interpolation
+  (via the `MET_NP` and `MET_P` control parameters).
 
-You will need wget (https://www.gnu.org/software/wget) to retrieve the
-data, and the netCDF Operators (NCO, https://nco.sourceforge.net) to
-convert the netCDF files for MPTRAC.
+Requirements:
+  - `wget` (https://www.gnu.org/software/wget) to retrieve data.
+  - NetCDF Operators (NCO, https://nco.sourceforge.net) for file conversion.
+  - A NASA EarthData account (https://urs.earthdata.nasa.gov) linked to GES DISC.
+  - Configuration of `~/.netrc` and `~/.urs_cookies` for authenticated wget access:
+    See https://disc.gsfc.nasa.gov/data-access#windows_wget for setup instructions.
 
-You will need a NASA EarthData account and link GES DISC to your
-account (https://urs.earthdata.nasa.gov) to access the MERRA-2
-data. You will need to configure the ~/.netrc and ~/.urs_cookies files
-to use wget to retrieve MERRA-2 data from GES DISC
-(https://disc.gsfc.nasa.gov/data-access#windows_wget).
+License:
+  MERRA-2 data are provided by NASA under the Earth Science Data and Information System (ESDIS) policy.
 EOF
     exit
 fi
