### 2025-06-23 #8177 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
