### 2025-06-23 5bb85ea
Added settings for meteo data.
