+++ projects/web_runner/app.py
@@ -222,4 +222,4 @@ def terms():
     return render_template('terms.html')
 
 if __name__ == '__main__':
-    app.run(host="0.0.0.0", port=7000, debug=True)
+    app.run(debug=True)
