### 2025-06-24 7e4516c
Update map plot. Added meteo variables to output.
