### 2025-06-24 f87a668
Update map plot.
