### 2025-06-24 c952a1c
Added more options for output.
