+++ projects/web_runner/app.py
@@ -8,7 +8,7 @@ import matplotlib.pyplot as plt
 import cartopy.crs as ccrs
 import cartopy.feature as cfeature
 
-# === Config ===
+# Config...
 ATM_INIT_CMD = '../../src/atm_init'
 TRAC_CMD = '../../src/trac'
 RUNS_DIR, ZIPS_DIR = 'runs/working', 'runs/zips'
@@ -17,7 +17,7 @@ os.makedirs(ZIPS_DIR, exist_ok=True)
 process_semaphore = threading.Semaphore(3)
 app = Flask(__name__)
 
-# === Meteo options ===
+# Meteo options...
 MET_OPTIONS = {
     'era5low_6h': {
         'METBASE': '/mnt/slmet-mnt/met_data/ecmwf/era5.1/resolution_1x1/nc/YYYY/MM/era5',
@@ -51,9 +51,11 @@ MET_OPTIONS = {
     }
 }
 
+# Clean up working directory...
 def delayed_cleanup(path, delay=600):
     threading.Thread(target=lambda: (time.sleep(delay), shutil.rmtree(path, ignore_errors=True)), daemon=True).start()
 
+# Execute shell command and check result...
 def run_command(cmd, timeout=300):
     if not process_semaphore.acquire(timeout=5):
         return -999, "❌ Server is too busy."
@@ -65,10 +67,12 @@ def run_command(cmd, timeout=300):
     finally:
         process_semaphore.release()
 
+# Calculate second since 2000-01-01, 00:00 UTC...
 def seconds_since_2000(time_str):
     dt = datetime.strptime(time_str, "%Y-%m-%d %H:%M:%S").replace(tzinfo=timezone.utc)
     return (dt - datetime(2000, 1, 1, tzinfo=timezone.utc)).total_seconds()
 
+# Create map plot of air parcel data...
 def create_plot(lons, lats, heights, projection='cartesian', region='global', central_lon=0.0, central_lat=0.0):
     import matplotlib.pyplot as plt
     import cartopy.crs as ccrs
@@ -106,6 +110,7 @@ def create_plot(lons, lats, heights, projection='cartesian', region='global', ce
     plt.close()
     return base64.b64encode(buf.getvalue()).decode()
 
+# Exit with error message...
 def validation_error(message):
     return render_template('error.html', stdout=f"❌ {message}"), 400
 
@@ -119,6 +124,7 @@ def download(run_id):
 @app.route('/run', methods=['POST'])
 def run():
     try:
+        # Get control parameters from input form...
         f = request.form
         to_f = lambda x: float(f[x])
         met_source = f.get('met_source', 'erai_6h')
@@ -141,6 +147,7 @@ def run():
         central_lon = to_f('central_lon') if 'central_lon' in f else 0.0
         central_lat = to_f('central_lat') if 'central_lat' in f else 0.0
 
+        # Check values...
         if abs(start_time - stop_time) > 30*86400: return validation_error("Duration exceeds 30 days.")
         if not (-100 <= z0 <= 100 and -100 <= z1 <= 100): return validation_error("Height range invalid.")
         if not (0 < dz <= 100): return validation_error("dz must be > 0 and ≤ 100 km.")
@@ -152,18 +159,24 @@ def run():
         if any(p < 0 for p in [ulon, ulat, uz, slon, slat, sz] + list(turb.values())): return validation_error("Negative values not allowed.")
         if atm_dt_out <= 0: return validation_error("Output frequency must be > 0.")
 
+        # Check number of air parcels...
         nz, nlat, nlon = map(lambda x: int(round(x[0]/x[1])) + 1, [(z1-z0, dz), (lat1-lat0, dlat), (lon1-lon0, dlon)])
[…]
