+++ projects/web_runner/app.py
@@ -67,38 +67,29 @@ def run_command(cmd, timeout=300):
     finally:
         process_semaphore.release()
 
-# Calculate second since 2000-01-01, 00:00 UTC...
+# Calculate seconds since 2000-01-01, 00:00 UTC...
 def seconds_since_2000(time_str):
-    dt = datetime.strptime(time_str, "%Y-%m-%d %H:%M:%S").replace(tzinfo=timezone.utc)
+    dt = datetime.strptime(time_str, "%Y-%m-%d %H:%M").replace(tzinfo=timezone.utc)
     return (dt - datetime(2000, 1, 1, tzinfo=timezone.utc)).total_seconds()
 
 # Create map plot of air parcel data...
 def create_plot(lons, lats, heights, projection='cartesian', region='global', central_lon=0.0, central_lat=0.0):
-    import matplotlib.pyplot as plt
-    import cartopy.crs as ccrs
-    import cartopy.feature as cfeature
-    import io, base64
-    if projection == 'cartesian':
-        proj = ccrs.PlateCarree()
-    elif projection == 'orthographic':
-        proj = ccrs.Orthographic(central_longitude=central_lon, central_latitude=central_lat)
-    elif projection == 'robinson':
-        proj = ccrs.Robinson(central_longitude=central_lon)
-    fig = plt.figure(figsize=(15, 12))
-    ax = plt.axes(projection=proj)
+    proj_map = {
+        'cartesian': ccrs.PlateCarree(),
+        'orthographic': ccrs.Orthographic(central_longitude=central_lon, central_latitude=central_lat),
+        'robinson': ccrs.Robinson(central_longitude=central_lon)
+    }
+    proj = proj_map.get(projection, ccrs.PlateCarree())
+    fig, ax = plt.subplots(figsize=(15, 12), subplot_kw={'projection': proj})
     if region == 'global':
         ax.set_global()
     ax.coastlines(color='gray', resolution='10m')
     ax.add_feature(cfeature.BORDERS, linewidth=0.5, edgecolor='gray')
-    sc = ax.scatter(
-        lons, lats, c=heights,
-        cmap='tab20c', s=10, edgecolors='none',
-        transform=ccrs.PlateCarree()
-    )
+    sc = ax.scatter(lons, lats, c=heights, cmap='tab20c', s=10, edgecolors='none', transform=ccrs.PlateCarree())
     gl = ax.gridlines(draw_labels=True, color='gray', linestyle='--', alpha=0.5)
     gl.xlabel_style = gl.ylabel_style = {'color': 'black'}
-    ax.set_facecolor('white')
-    fig.patch.set_facecolor('white')
+    for obj in [ax, fig]:
+        obj.set_facecolor('white')
     cbar = plt.colorbar(sc, orientation='horizontal', pad=0.05, aspect=50)
     cbar.set_label('Log-pressure height [km]', color='black')
     cbar.ax.xaxis.set_tick_params(color='black')
@@ -141,11 +132,11 @@ def run():
         ulon, ulat, uz = to_f('ulon'), to_f('ulat'), to_f('uz')
         slon, slat, sz = to_f('slon'), to_f('slat'), to_f('sz')
         turb = {k: to_f(k) for k in ['turb_dx_pbl','turb_dx_trop','turb_dx_strat','turb_dz_pbl','turb_dz_trop','turb_dz_strat','turb_mesox','turb_mesoz']}
-        atm_dt_out = to_f('atm_dt_out') if 'atm_dt_out' in f else 3600
+        atm_dt_out = to_f('atm_dt_out')
         plot_region = f.get('plot_region', 'global')
         map_projection = f.get('map_projection', 'cartesian')
-        central_lon = to_f('central_lon') if 'central_lon' in f else 0.0
[…]
