+++ docs/joss/paper.bib
@@ -28,7 +28,7 @@
 }
 
 @article{draxler91,
-  title =	 {Across North America tracer experiment (ANATEX):
+  title =	 {{Across North America tracer experiment} (ANATEX):
                   Sampling and analysis},
   journal =	 {Atmospheric Environment. Part A. General Topics},
   volume =	 25,
@@ -54,7 +54,7 @@
                   and Michele Rienecker and Siegfried D. Schubert and
                   Meta Sienkiewicz and Bin Zhao",
   title =	 "The {Modern-Era Retrospective Analysis for Research
-                  and Applications}, Version 2 (MERRA-2)",
+                  and Applications}, {Version 2} {(MERRA-2)}",
   journal =	 "Journal of Climate",
   year =	 2017,
   volume =	 30,
@@ -123,9 +123,9 @@
                   Heng, Y. and Liu, M. and Haghighi Mood, K. and
                   Stein, O. and Thomas, N. and Vogel, B. and Wu,
                   X. and Zou, L.},
-  TITLE =	 {Massive-Parallel Trajectory Calculations version 2.2
-                  (MPTRAC-2.2): Lagrangian transport simulations on
-                  graphics processing units (GPUs)},
+  TITLE =	 {{Massive-Parallel Trajectory Calculations version
+                  2.2 (MPTRAC-2.2)}: Lagrangian transport simulations
+                  on graphics processing units {(GPUs)}},
   JOURNAL =	 {Geoscientific Model Development},
   VOLUME =	 15,
   YEAR =	 2022,
@@ -166,7 +166,7 @@
 
 @misc{hoffmann25,
   author =	 {Hoffmann, L.},
-  title =	 {MPTRAC Convective Transport Animation},
+  title =	 {{MPTRAC} Convective Transport Animation},
   month =	 apr,
   year =	 2025,
   publisher =	 {Zenodo},
@@ -250,7 +250,7 @@
                   Z. and Heng, Y. and Wu, X.},
   TITLE =	 {Improved representation of volcanic sulfur dioxide
                   depletion in Lagrangian transport simulations: a
-                  case study with MPTRAC v2.4},
+                  case study with {MPTRAC} v2.4},
   JOURNAL =	 {Geoscientific Model Development},
   VOLUME =	 16,
   YEAR =	 2023,
@@ -347,8 +347,9 @@
 }
 
 @article{vandop98,
-  title =	 {ETEX: A European tracer experiment; observations,
-                  dispersion modelling and emergency response},
+  title =	 {{ETEX}: A {European} tracer experiment;
+                  observations, dispersion modelling and emergency
+                  response},
   journal =	 {Atmospheric Environment},
   volume =	 32,
   number =	 24,
