### 2025-06-25 #8177 @lars2015
Dear Kristen Thyng [@kthyng],

Thank you for your feedback.

We've checked and corrected the capitalization in the references in the .bib file where necessary.

We also checked the metadata in the Zenodo archive and updated the affiliations and DOIs for three of our co-authors. Everything now appears to be consistent between the JOSS paper and the Zenodo record.

Please let us know if anything else needs adjustment.

Best regards
Lars Hoffmann
