+++ src/mptrac.h
@@ -159,6 +159,10 @@
 #include "chem_Sparse.h"
 #endif
 
+#ifdef ECCODES
+#include "eccodes.h"
+#endif
+
 /* ------------------------------------------------------------
    Constants...
    ------------------------------------------------------------ */
@@ -620,6 +624,71 @@
  */
 #define DOTP(a, b)				\
   (a[0]*b[0]+a[1]*b[1]+a[2]*b[2])
+
+/**
+ * @brief Execute a ECCODES command and check for errors.
+ *
+ * This macro executes a ECCODES command and checks the result. If the
+ * result indicates an error, it prints the error message using
+ * ERRMSG.
+ *
+ * @param cmd ECCODES command to execute.
+ *
+ * @author Nils Nobre Wittwer
+ */
+#define ECC(cmd) {							\
+    int ecc_result=(cmd);						\
+    if(ecc_result!=0)							\
+      ERRMSG("ECCODES error: %s", codes_get_error_message(ecc_result));	\
+  }
+
+/**
+ * @brief Writes 2-D data from a grib message into the met struct.
+ *
+ * This Macro writes 2-D data from an one dimensional grib message into the
+ * corresponding 2-D variable in the met struct. 
+ * 
+ * @param variable Name of the current meterological variable 
+ * @param target Pointer to the 2-D array in the met struct where the data will be stored.
+ * @param scaling_factor Scaling factor to apply to the data.
+ * @param found_flag Flag to store, that the variable was found in the grib message.
+ *
+ * @author Nils Nobre Wittwer
+ */
+#define ECC_READ_2D(variable,target,scaling_factor,found_flag){	\
+  if( strcmp(short_name,variable)==0){\
+    for (int ix = 0; ix < met->nx; ix++) {\
+          for (int iy = 0; iy < met->ny; iy++) {\
+              target[ix][iy] = (float)(values[iy * met->nx + ix]*scaling_factor);\
+          }\
+      }\
+    found_flag =1;\
+  }\
+}
+
+/**
+ * @brief Writes 3D data from a grib message into the met struct.
+ *
+ * This Macro writes 3D data from an one dimensional grib message into the
+ * corresponding 3D variable in the met struct. 
+ * 
+ * @param variable Name of the current meterological variable.
+ * @param target Pointer to the 3D array in the met struct where the data will be stored.
+ * @param scaling_factor Scaling factor to apply to the data.
+ * @param found_flag Counter to store, how many messages countaining data for this variable have been read.
+ *
+ * @author Nils Nobre Wittwer
+ */
+#define ECC_READ_3D(variable,level,target,scaling_factor,found_flag){\
+  if( strcmp(short_name,variable)==0){\
+    for (int ix = 0; ix < met->nx; ix++) {\
+          for (int iy = 0; iy < met->ny; iy++) {\
+              target[ix][iy][level] = (float) (values[iy * met->nx + ix]*scaling_factor);\
+          }\
+      }\
+    found_flag +=1;\
+  }\
+}
 
 /**
  * @brief Calculate the floating-point remainder of dividing x by y.
@@ -6982,6 +7051,66 @@ void read_met_geopot(
   met_t * met);
 
 /**
+ * @brief Reads global meteorological information from a grib file.
+ *
+ * This function reads meteorological grid information from a Grib
+ * file, including time and spatial dimensions. The function computes the 
+ * latitude and longitude grid based on the provided boundaries and icnrement values.
+ *
+ * @param handles A pointer to an array of codes_handle pointers representing the Grib messages.
+ * @param count_handles The total number of Grib messages in the handles array.
+ * @param met A pointer to a structure to store meteorological data.
+ *
+ * The function performs the following steps:
+ * - Sets up a timer to monitor the reading time for meteorological grid information.
+ * - Determines the time information from the data file.
[…]
