+++ tests/grib_test/run.sh
@@ -12,20 +12,16 @@ trac=../../src
 # Create directories...
 rm -rf data && mkdir -p data
 
-$trac/met_zm - ./data/testzm.tab\
-		./gribdata/gb_2011060600_small_XX.grb\
-		MET_TYPE 6 \
-		MET_PRESS_LEVEL_DEF 6 
+# Write zonal mean data...
+$trac/met_zm - ./data/testzm.tab \
+	     ./gribdata/gb_2011060600_small_XX.grb \
+	     MET_TYPE 6 MET_PRESS_LEVEL_DEF 6 
 
-$trac/met_map - ./data/testmap.tab\
-		./gribdata/gb_2011060600_small_XX.grb\
-		MET_TYPE 6 \
-		MET_PRESS_LEVEL_DEF 6\
-		MAP_LON0 5.1\
-		MAP_LON1 15\ 
-		MAP_LAT0 45\
-		MAP_LAT1 54.9\
-
+# Write map data...
+$trac/met_map - ./data/testmap.tab \
+	      ./gribdata/gb_2011060600_small_XX.grb \
+	      MET_TYPE 6 MET_PRESS_LEVEL_DEF 6 \
+	      MAP_LON0 5.1 MAP_LON1 15
 
 # Compare files...
 echo -e "\nCompare results..."
