+++ src/Makefile
@@ -272,7 +272,7 @@ wrapper: trac_fortran
 
 check: $(TESTS)
 
-$(TESTS) cms_test gpu_test kpp_test wrapper_test: all
+$(TESTS) cms_test gpu_test grib_test kpp_test wrapper_test: all
 	@(echo "\n===== Running \"$@\" ... =====") ; \
 	  cd ../tests/$@ ; ./run.sh \
 	  && (echo "\n===== Test \"$@\" passed! =====") \
