### 2025-06-29 607a836
Fixed code formatting and comments.
