+++ src/mptrac.h
@@ -655,16 +655,16 @@
  *
  * @author Nils Nobre Wittwer
  */
-#define ECC_READ_2D(variable,target,scaling_factor,found_flag){	\
-  if( strcmp(short_name,variable)==0){\
-    for (int ix = 0; ix < met->nx; ix++) {\
-          for (int iy = 0; iy < met->ny; iy++) {\
-              target[ix][iy] = (float)(values[iy * met->nx + ix]*scaling_factor);\
-          }\
-      }\
-    found_flag =1;\
-  }\
-}
+#define ECC_READ_2D(variable,target,scaling_factor,found_flag){		\
+    if( strcmp(short_name,variable)==0){				\
+      for (int ix = 0; ix < met->nx; ix++) {				\
+	for (int iy = 0; iy < met->ny; iy++) {				\
+	  target[ix][iy] = (float)(values[iy * met->nx + ix]*scaling_factor); \
+	}								\
+      }									\
+      found_flag =1;							\
+    }									\
+  }
 
 /**
  * @brief Writes 3D data from a grib message into the met struct.
@@ -679,16 +679,16 @@
  *
  * @author Nils Nobre Wittwer
  */
-#define ECC_READ_3D(variable,level,target,scaling_factor,found_flag){\
-  if( strcmp(short_name,variable)==0){\
-    for (int ix = 0; ix < met->nx; ix++) {\
-          for (int iy = 0; iy < met->ny; iy++) {\
-              target[ix][iy][level] = (float) (values[iy * met->nx + ix]*scaling_factor);\
-          }\
-      }\
-    found_flag +=1;\
-  }\
-}
+#define ECC_READ_3D(variable,level,target,scaling_factor,found_flag){	\
+    if( strcmp(short_name,variable)==0){				\
+      for (int ix = 0; ix < met->nx; ix++) {				\
+	for (int iy = 0; iy < met->ny; iy++) {				\
+	  target[ix][iy][level] = (float) (values[iy * met->nx + ix]*scaling_factor); \
+	}								\
+      }									\
+      found_flag +=1;							\
+    }									\
+  }
 
 /**
  * @brief Calculate the floating-point remainder of dividing x by y.
@@ -7053,19 +7053,19 @@ void read_met_geopot(
 /**
  * @brief Reads global meteorological information from a grib file.
  *
- * This function reads meteorological grid information from a Grib
+ * This function reads meteorological grid information from a grib
  * file, including time and spatial dimensions. The function computes the 
  * latitude and longitude grid based on the provided boundaries and icnrement values.
  *
- * @param handles A pointer to an array of codes_handle pointers representing the Grib messages.
- * @param count_handles The total number of Grib messages in the handles array.
+ * @param handles A pointer to an array of codes_handle pointers representing the grib messages.
+ * @param count_handles The total number of grib messages in the handles array.
  * @param met A pointer to a structure to store meteorological data.
  *
  * The function performs the following steps:
  * - Sets up a timer to monitor the reading time for meteorological grid information.
  * - Determines the time information from the data file.
- * - Retrieves grid dimensions (longitude, latitude, and vertical levels) from the Grib file.
- * - Reads longitudes and latitudes boundaries and increments from the Grib file and computes the grid.
+ * - Retrieves grid dimensions (longitude, latitude, and vertical levels) from the grib file.
+ * - Reads longitudes and latitudes boundaries and increments from the grib file and computes the grid.
  * - Logs the retrieved grid information for verification and debugging purposes.
  *
  *
@@ -7073,20 +7073,20 @@ void read_met_geopot(
  */
 #ifdef ECCODES
 void read_met_global_grib(
-  codes_handle** handles,
+  codes_handle ** handles,
[…]
