+++ src/mptrac.c
@@ -5736,14 +5736,13 @@ int mptrac_read_met(
       if (read_met_bin(filename, ctl, met) != 1)
 	return 0;
     }
-
-    #ifdef ECCODES
+#ifdef ECCODES
     /* Read grib data... */
-    else if (ctl->met_type == 6){
+    else if (ctl->met_type == 6) {
       if (read_met_grib(filename, ctl, clim, met) != 1)
-  return 0;
+	return 0;
     }
-    #endif
+#endif
 
     /* Not implemented... */
     else
@@ -7608,7 +7607,7 @@ void read_met_grid(
 
 #ifdef ECCODES
 void read_met_global_grib(
-  codes_handle** handles,
+  codes_handle **handles,
   int count_handles,
   met_t *met) {
 
@@ -7616,97 +7615,100 @@ void read_met_global_grib(
   SELECT_TIMER("READ_MET_GLOBAL_GRIB", "INPUT", NVTX_READ);
   LOG(2, "Read meteo grid information...");
 
-  
-  /*Read date...*/
+  /*Read date... */
   char datestr[50];
   char timestr[50];
-  char year[20],month[20],day[20],hour[20];
+  char year[20], month[20], day[20], hour[20];
   size_t s = sizeof(datestr);
 
-
-  ECC(codes_get_string(handles[0],"dataDate",datestr,&s));
-  ECC(codes_get_string(handles[0],"dataTime",timestr,&s));
-  strncpy(year,datestr,4);
+  ECC(codes_get_string(handles[0], "dataDate", datestr, &s));
+  ECC(codes_get_string(handles[0], "dataTime", timestr, &s));
+  strncpy(year, datestr, 4);
   year[4] = '\0';
-  strncpy(month,datestr+4,2);
+  strncpy(month, datestr + 4, 2);
   month[2] = '\0';
-  strncpy(day,datestr+6,2);
+  strncpy(day, datestr + 6, 2);
   day[2] = '\0';
-  strncpy(hour,timestr,2);
+  strncpy(hour, timestr, 2);
   hour[2] = '\0';
-  time2jsec(atoi(year),atoi(month),atoi(day),atoi(hour),0,0,0,&(met->time));
+  time2jsec(atoi(year), atoi(month), atoi(day), atoi(hour), 0, 0, 0,
+	    &(met->time));
 
   /* Write info... */
   LOG(2, "Time: %.2f (%d-%02d-%02d, %02d:%02d UTC)",
-      met->time, atoi(year),atoi(month),atoi(day),atoi(hour), 0);
-  
+      met->time, atoi(year), atoi(month), atoi(day), atoi(hour), 0);
+
   /* Read grid information... */
-  long count_lat = 0,count_lon= 0;
-  ECC(codes_get_long(handles[0],"Nj",&count_lat));
-  ECC(codes_get_long(handles[0],"Ni",&count_lon));
+  long count_lat = 0, count_lon = 0;
+  ECC(codes_get_long(handles[0], "Nj", &count_lat));
+  ECC(codes_get_long(handles[0], "Ni", &count_lon));
   met->ny = (int) count_lat;
   met->nx = (int) count_lon;
 
   /* Check grid dimensions... */
   LOG(2, "Number of longitudes: %d", met->nx);
-  if(met->nx<2 || met->nx>EX)
+  if (met->nx < 2 || met->nx > EX)
     ERRMSG("Number of longitudes out of range!");
   LOG(2, "Number of latitudes: %d", met->ny);
-  if(met->ny<2 || met->ny>EY)
-    ERRMSG("Number of latitudes out of range!");    
-  
-  double first_lon,last_lon,first_lat,last_lat,inc_lon,inc_lat;
[…]
