### 2025-06-29 59fdb0a
Update DEPENDENCIES.md
