### 2025-07-01 807df2e
A preliminary GPU version
