+++ src/mptrac.c
@@ -3837,6 +3837,9 @@ void module_sort(
   double *restrict const a = (double *) malloc((size_t) np * sizeof(double));
   int *restrict const p = (int *) malloc((size_t) np * sizeof(int));
   double amax = (met0->nx*met0->ny + met0->ny)*met0->np + met0->np;
+  
+  /* MPI information... */
+  //MPI_Comm_rank(MPI_COMM_WORLD, &rank);
 
 #ifdef _OPENACC
 #pragma acc enter data create(a[0:np],p[0:np],amax)
@@ -3850,13 +3853,14 @@ void module_sort(
 #pragma omp parallel for default(shared)
 #endif
   for (int ip = 0; ip < np; ip++) {
-    if ((int) atm->q[ctl->qnt_subdomain][ip] != -1)
-      a[ip] = 0;
-      /*(double) ((locate_reg(met0->lon, met0->nx, atm->lon[ip]) * met0->ny +
+    if ((int) atm->q[ctl->qnt_subdomain][ip] != -1) {
+      a[ip] = (double) ((locate_reg(met0->lon, met0->nx, atm->lon[ip]) * met0->ny +
 		 locate_irr(met0->lat, met0->ny, atm->lat[ip]))
-		* met0->np + locate_irr(met0->p, met0->np, atm->p[ip]));*/
-    else
+		* met0->np + locate_irr(met0->p, met0->np, atm->p[ip]));
+      // Alos if destination != rank amax+1, then below amax+2
+     } else {
       a[ip] = amax+1;
+     }
     p[ip] = ip;
   }
   
@@ -5580,60 +5584,65 @@ void mptrac_run_timestep(
   MPI_Comm_rank(MPI_COMM_WORLD,&rankd);
 
   /* Domain decomposition... */
-#pragma acc update host(atm, cache)
   if (ctl->dd_subdomains_meridional*ctl->dd_subdomains_zonal > 1) {
     
-    
     /* Initialize particles locally... */
-    int nparticles = 0;
+    int nparticles = 0; 
     int neighbours[NNMAX];
     particle_t* particles;
     ALLOC(particles, particle_t, NP);
 
-    /* Get the MPI information... */
+    /* Get the MPI information on CPU... */
     int rank, size;
-        
     MPI_Barrier(MPI_COMM_WORLD);
     MPI_Comm_rank(MPI_COMM_WORLD, &rank);
-    size = ctl->dd_subdomains_meridional*ctl->dd_subdomains_zonal;
-   
-    //LOG(1, "Starting module_dd with %d particles on rank %d", atm->np, rank);
-            
+    size = ctl->dd_subdomains_meridional*ctl->dd_subdomains_zonal;  
+    
     /* Define grid neighbours ... */
-    //LOG(1, "Define grid neighbours.");
+    // TODO: This actually only needs to be done ones!
     dd_get_rect_neighbour(*ctl, neighbours, rank, size);
-        
+            
+    /* Create data region... */    
+#pragma acc enter data create( nparticles, neighbours[:NNMAX], particles[:NP], rank, size)
+   
+    /* Update GPU... */
+#pragma acc update device( rank, size, nparticles, neighbours[:NNMAX])
+
     /* Assign particles to new subdomains... */
-    //LOG(1, "Assign particles to new subdomains.")
-    dd_assign_rect_subdomains_atm( atm, *met0, *ctl, rank, neighbours, 0);
-    
+    dd_assign_rect_subdomains_atm( atm, *met0, ctl, rank, neighbours, 0);
+            
     /* Transform from struct of array to array of struct... */
-    //LOG(1, "Transform from SoA to AoS.")
-    //atm2particles(atm, particles, *ctl)
-    atm2particles( atm, particles, *ctl, &nparticles, cache);
-        
[…]
