### 2025-07-03 7fce6f5
Fix library build script for eccodes.
