+++ libs/build.sh
@@ -15,6 +15,7 @@ ifBuildZSTD=false
 ifBuildHDF5=false
 ifBuildNETCDF=false
 ifBuildKPP=false
+ifBuildECCODES=false
 ifSkip=false
 
 numProcs=`nproc --all`
@@ -32,8 +33,7 @@ strFileZLIB="zlib-1.3.1"
 strFileSZIP="szip-2.1.1"
 strFileZSTD="zstd-1.5.5"
 strFileKPP="KPP"
-strFileEcCodes="eccodes-2.38.3"
-
+strFileECCODES="eccodes-2.38.3"
 
 export LD_LIBRARY_PATH=$strBuildDir/lib:$LD_LIBRARY_PATH
 
@@ -62,7 +62,7 @@ do
 	d) ifBuildHDF5=true   ; echo "HDF5 is selected     " ;;
 	n) ifBuildNETCDF=true ; echo "NETCDF is selected   " ;;
 	k) ifBuildKPP=true    ; echo "KPP is selected      " ;;
-    e) ifBuildEcCodes=true; echo "ecCodes is selected  " ;;
+	e) ifBuildECCODES=true; echo "ECCODES is selected  " ;;
 	p) numProcs=${OPTARG}  
            numTotalProcs=`nproc --all`
            if [ $numProcs -lt $numTotalProcs ]; then
@@ -90,7 +90,7 @@ do
 	   printf -- "-d         : build HDF5   (prerequisites: CURL, ZLIB, SZIP)  \n" 
 	   printf -- "-n         : build NETCDF (prerequisite: HDF5)  \n" 
 	   printf -- "-k         : build KPP    \n" 
-       printf -- "-e         : build ecCodes\n"
+	   printf -- "-e         : build ECCODES\n"
 	   printf -- "-p [cores] : how many [cores] to be used by make -j [cores]   \n" 
 	   printf -- "           : default is maximum number of cores               \n" 
 	   printf -- "--------------------------------------------------------------\n"	
@@ -261,16 +261,15 @@ if [ $ifBuildAll = true ] || [ $ifBuildKPP = true ] ; then
 fi
 
 # ecCodes...
-if [ $ifBuildAll = true ] || [ $ifBuildEcCodes = true ] ; then
+if [ $ifBuildAll = true ] || [ $ifBuildECCODES = true ] ; then
     cd $strLibsDir
     printf "starting to compile ecCodes...\n"
-    strTarget=$strFileEcCodes
+    strTarget=$strFileECCODES
     cp $strTarget.tar.gz $strBuildDir/src && cd $strBuildDir/src && tar xvf $strTarget.tar.gz
     cd $strBuildDir/src && cmake -DCMAKE_INSTALL_PREFIX=$strBuildDir/src -DBUILD_SHARED_LIBS=BOTH -DENABLE_AEC=OFF -DENABLE_JPEG=OFF -DENABLE_FORTRAN=OFF -DENABLE_NETCDF=OFF $strTarget && make -j $numProcs && ctest -j $numProcs && make install \
         && cp -a include/* $strBuildDir/include/ \
         && cp -a lib/libeccodes* $strBuildDir/lib/ \
         || exit
-
 fi
 
 # Finish...
