+++ doxygen/search/functions_c.js
@@ -1,0 +1,39 @@
+var searchData=
+[
+  ['read_5fatm_5fasc_0',['read_atm_asc',['../mptrac_8h.html#af0092541e8eb2c51fc9a3fcd98043f3f',1,'read_atm_asc(const char *filename, const ctl_t *ctl, atm_t *atm):&#160;mptrac.c'],['../mptrac_8c.html#af0092541e8eb2c51fc9a3fcd98043f3f',1,'read_atm_asc(const char *filename, const ctl_t *ctl, atm_t *atm):&#160;mptrac.c']]],
+  ['read_5fatm_5fbin_1',['read_atm_bin',['../mptrac_8h.html#ac51355e97453a01a70446623d0c382e5',1,'read_atm_bin(const char *filename, const ctl_t *ctl, atm_t *atm):&#160;mptrac.c'],['../mptrac_8c.html#ac51355e97453a01a70446623d0c382e5',1,'read_atm_bin(const char *filename, const ctl_t *ctl, atm_t *atm):&#160;mptrac.c']]],
+  ['read_5fatm_5fclams_2',['read_atm_clams',['../mptrac_8c.html#ad55d18eefcfa90b7c40d4a37dbb8fbf0',1,'read_atm_clams(const char *filename, const ctl_t *ctl, atm_t *atm):&#160;mptrac.c'],['../mptrac_8h.html#ad55d18eefcfa90b7c40d4a37dbb8fbf0',1,'read_atm_clams(const char *filename, const ctl_t *ctl, atm_t *atm):&#160;mptrac.c']]],
+  ['read_5fatm_5fnc_3',['read_atm_nc',['../mptrac_8c.html#a2569c3ce3d8d46ce8aa186d42685494c',1,'read_atm_nc(const char *filename, const ctl_t *ctl, atm_t *atm):&#160;mptrac.c'],['../mptrac_8h.html#a2569c3ce3d8d46ce8aa186d42685494c',1,'read_atm_nc(const char *filename, const ctl_t *ctl, atm_t *atm):&#160;mptrac.c']]],
+  ['read_5fclim_5fphoto_4',['read_clim_photo',['../mptrac_8c.html#a1a501ec43c01f1d742a133a7a3289d59',1,'read_clim_photo(const char *filename, clim_photo_t *photo):&#160;mptrac.c'],['../mptrac_8h.html#a1a501ec43c01f1d742a133a7a3289d59',1,'read_clim_photo(const char *filename, clim_photo_t *photo):&#160;mptrac.c']]],
+  ['read_5fclim_5fphoto_5fhelp_5',['read_clim_photo_help',['../mptrac_8c.html#a4392a0ab24a984ac30ed1ef85a89257d',1,'read_clim_photo_help(const int ncid, const char *varname, const clim_photo_t *photo, double var[CP][CSZA][CO3]):&#160;mptrac.c'],['../mptrac_8h.html#a4392a0ab24a984ac30ed1ef85a89257d',1,'read_clim_photo_help(const int ncid, const char *varname, const clim_photo_t *photo, double var[CP][CSZA][CO3]):&#160;mptrac.c']]],
+  ['read_5fclim_5fts_6',['read_clim_ts',['../mptrac_8c.html#a3510fe56a6a6e589b5c2474495af03e2',1,'read_clim_ts(const char *filename, clim_ts_t *ts):&#160;mptrac.c'],['../mptrac_8h.html#a3510fe56a6a6e589b5c2474495af03e2',1,'read_clim_ts(const char *filename, clim_ts_t *ts):&#160;mptrac.c']]],
[…]

[…]
