### 2025-07-08 e788b9b
Added NCEP GFS forecasts.
