+++ projects/web_runner/app.py
@@ -48,6 +48,11 @@ MET_OPTIONS = {
         'METBASE': '/mnt/slmet-mnt/met_data/ncep/reanalysis2/nc/YYYY/ncep2',
         'DT_MET': 21600.0,
         'MET_PRESS_LEVEL_DEF': -1
+    },
+    'gfs_3h': {
+        'METBASE': '/p/data1/slmet/met_data/ncep/gfs/data_YYYY_MM_DD/gfs',
+        'DT_MET': 10800.0,
+        'MET_PRESS_LEVEL_DEF': -1
     }
 }
 
