### 2025-07-08 d162ea7
Fixed path for NCEP GFS forecasts.
