+++ README.md
@@ -15,7 +15,7 @@ Massive-Parallel Trajectory Calculations (MPTRAC) is a Lagrangian particle dispe
 [![tests](https://img.shields.io/github/actions/workflow/status/slcs-jsc/mptrac/tests.yml?branch=master&label=tests)](https://github.com/slcs-jsc/mptrac/actions)
 [![docs](https://img.shields.io/github/actions/workflow/status/slcs-jsc/mptrac/docs.yml?branch=master&label=docs)](https://slcs-jsc.github.io/mptrac)
 [![status](https://joss.theoj.org/papers/dc49cd7be52f2a47c41cc20143f3719a/status.svg)](https://joss.theoj.org/papers/dc49cd7be52f2a47c41cc20143f3719a)
-[![MPTRAC Web Runner](https://img.shields.io/badge/Run%20MPTRAC%20Online-Web%20Runner-blue?logo=launchpad&style=flat-square)](https://mptrac.jsc.fz-juelich.de)
+[![MPTRAC Web Runner](https://img.shields.io/badge/Web%20Runner-online-green?logo=launchpad)](http://mptrac.jsc.fz-juelich.de)
 [![license](https://img.shields.io/github/license/slcs-jsc/mptrac.svg)](https://github.com/slcs-jsc/mptrac/blob/master/COPYING)
 [![doi](https://zenodo.org/badge/DOI/10.5281/zenodo.4400597.svg)](https://doi.org/10.5281/zenodo.4400597)
 
