### 2025-07-08 addad89
Another fix of NCEP GFS path.
