+++ projects/web_runner/app.py
@@ -19,41 +19,20 @@ app = Flask(__name__)
 
 # Meteo options...
 MET_OPTIONS = {
-    'era5low_6h': {
-        'METBASE': '/mnt/slmet-mnt/met_data/ecmwf/era5.1/resolution_1x1/nc/YYYY/MM/era5',
-        'DT_MET': 21600.0,
-        'MET_PRESS_LEVEL_DEF': 6
-    },
-    'erai_6h': {
-        'METBASE': '/mnt/slmet-mnt/met_data/ecmwf/era_interim/pressure_0.75deg_v2/nc/YYYY/ei',
-        'DT_MET': 21600.0,
-        'MET_PRESS_LEVEL_DEF': -1
-    },
-    'merra2_3h': {
-        'METBASE': '/mnt/slmet-mnt/met_data/nasa/merra-2/hybrid/YYYY/merra2',
-        'DT_MET': 10800.0,
-        'MET_PRESS_LEVEL_DEF': 6
-    },
-    'merra2_6h': {
-        'METBASE': '/mnt/slmet-mnt/met_data/nasa/merra-2/hybrid/YYYY/merra2',
-        'DT_MET': 21600.0,
-        'MET_PRESS_LEVEL_DEF': 6
-    },
-    'ncep_6h': {
-        'METBASE': '/mnt/slmet-mnt/met_data/ncep/reanalysis/nc/YYYY/ncep',
-        'DT_MET': 21600.0,
-        'MET_PRESS_LEVEL_DEF': -1
-    },
-    'ncep2_6h': {
-        'METBASE': '/mnt/slmet-mnt/met_data/ncep/reanalysis2/nc/YYYY/ncep2',
-        'DT_MET': 21600.0,
-        'MET_PRESS_LEVEL_DEF': -1
-    },
-    'gfs_3h': {
-        'METBASE': '/mnt/slmet-mnt/met_data/ncep/gfs/data_YYYY_MM_DD/gfs',
-        'DT_MET': 10800.0,
-        'MET_PRESS_LEVEL_DEF': -1
-    }
+    'era5low_6h': dict(METBASE='/mnt/slmet-mnt/met_data/ecmwf/era5.1/resolution_1x1/nc/YYYY/MM/era5',
+                       DT_MET=21600, MET_PRESS_LEVEL_DEF=6, MET_NAME='ERA5'),
+    'erai_6h': dict(METBASE='/mnt/slmet-mnt/met_data/ecmwf/era_interim/pressure_0.75deg_v2/nc/YYYY/ei',
+                    DT_MET=21600, MET_PRESS_LEVEL_DEF=-1, MET_NAME='ERA-Interim'),
+    'merra2_3h': dict(METBASE='/mnt/slmet-mnt/met_data/nasa/merra-2/hybrid/YYYY/merra2',
+                      DT_MET=10800, MET_PRESS_LEVEL_DEF=6, MET_NAME='MERRA-2'),
+    'merra2_6h': dict(METBASE='/mnt/slmet-mnt/met_data/nasa/merra-2/hybrid/YYYY/merra2',
+                      DT_MET=21600, MET_PRESS_LEVEL_DEF=6, MET_NAME='MERRA-2'),
+    'ncep_6h': dict(METBASE='/mnt/slmet-mnt/met_data/ncep/reanalysis/nc/YYYY/ncep',
+                    DT_MET=21600, MET_PRESS_LEVEL_DEF=-1, MET_NAME='NCEP-NCAR Reanalysis 1'),
+    'ncep2_6h': dict(METBASE='/mnt/slmet-mnt/met_data/ncep/reanalysis2/nc/YYYY/ncep2',
[…]
