### 2025-07-08 #8177 @kthyng
Congratulations on your new publication @lars2015! Many thanks to editor @AnjaliSandip and to reviewers @dostuffthatmatters, @KYANJO, and @filippi for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
